docs(l1): readme update and kurtosis services addition (#928)
**Motivation**

After walking through the README as someone new to the project
everything worked as expected except for the Hive test runs.

**Description**

This PR adds golang to the `.tool-versions` (commented, given that it's
not needed for the rest of the node execution) and a small prereq
section to the hive readme about installing the plugin and uncommenting
it in case of wanting to run the hive tests locally.

Also I added a couple of services to the kurtosis `network-params.yml`,
`el_forkmon` works as a monitor of the canonical chain and checks all
clients are synced between themselves and the amount of bad blocks
they've processed. On the other hand `tx|blob_spammer` adds more
sustance to the run, it mainly affects other nodes given that we don't
have p2p already implemented but will be picked up then.