feat(l1): initiate rlpx connections from discv4 (#936)
**Motivation**

DiscV4 protocol discovers new peers, and after deciding it is a valid
peer, a TCP connection is established and RLPx protocol starts

**Description**

Now, when receiving a DiscV4 Pong message and after evaluating the peer,
a RLPx connection is created and established. If handshake fails it
removes the peer from the Kademlia table.

Closes #837.

Also removed some hard-coded testing code.

Once the listen loop is built (#840) there may be other conditions where
a peer has to be discarded. (eg. when the other peer sends a Disconnect
message). After #840 is completed we may create some more issues on this
regard.

---------

Co-authored-by: Maximo Palopoli <96491141+maximopalopoli@users.noreply.github.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Federico Borello <156438142+fborello-lambda@users.noreply.github.com>
Co-authored-by: Marcos Nicolau <76252340+MarcosNicolau@users.noreply.github.com>
Co-authored-by: Rodrigo Oliveri <rodrigooliveri10@gmail.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: Manuel Iñaki Bilbao <manuel.bilbao@lambdaclass.com>
Co-authored-by: Martin Paulucci <martin.c.paulucci@gmail.com>
Co-authored-by: François <67512048+gsbujo@users.noreply.github.com>