feat(l1): enhance el assertoor ci test by cherrypicking more checks (#963)
**Motivation**

This PR tries to add as much test as possible to our assertoor CI run.

**Description**

The idea is to add every check that is useful for validating the el
execution without incurring in long running times to avoid making the CI
slower. Right now every useful check from the
[supported](https://github.com/ethpandaops/assertoor/wiki#supported-tasks-in-assertoor)
is added except for:
- ✅ `check_consensus_finality` | time: ~10m
- ✅ `check_consensus_attestation_stats` | time: ~10m

Apart from checks, we are not using most of the `generate_*` tasks due
to timing or errors:
- ✅ `run_transaction_test` |  time: ~10m
- ✅ `run_blob_transaction_test` | time: ~6m
- ❎ `run_opcodes_transaction_test` | time: ~5m (We have issues
processing some of the transactions)

For the already available assertor playbooks, [see
here](https://github.com/ethpandaops/assertoor-test/tree/master/assertoor-tests).

The `run_blob_transaction_test` is the only one added until now from
transaction checks and the total test take nearly 15m (which is less
than the largest hive test in the CI).

Closes #953