feat(l2): deposit verification (#1000)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Deposits information must be saved on L1 until it gets verified so we
can trust the proposer don't put any invalid mint transaction.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
When send the commit, the proposer will also send the deposit logs, a
hash with format `N || keccak(d1 || d2 || ... || dN)[2..32]`, with `N`
being the number of deposits executed encoded in 2 bytes, and `dX =
keccack(to, value)` of the transaction `X`. That info is stored on an
array until a `verify` call removes it.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #977