feat(levm): implement `Cache` and refactor `Db` (#991)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We currently have an in-memory database that should soon be replaced by
the node's in-disk database. For that purpose we want to allow us to
switch between both kinds of databases. The need to implement this PR's
features and refactor the `Db` arose while working on #904.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
This PR includes:
- Adding a `Cache` to store warm accounts. This removes the need of
having `accessed_accounts` and `accessed_storage_slots` sets in
`Substate` because we know that if they are cached then they are warm.
- Making our `Db` implement the `Database` trait and interact with it
only using methods and not it's attributes, so in the future we can
implement that trait for the actual node's database.
- Fix call opcodes and remove delegate attribute from `CallFrame`.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Part of #814.

---------

Co-authored-by: Juani Medone <juan.medone@lambdaclass.com>
Co-authored-by: maximopalopoli <mpalopoli@fi.uba.ar>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>