chore(l1): reduce unnecessary docker builds and remove duplicate workflow executions in main (#1032)
**Motivation**

We are re-building our docker image across the jobs/workflows multiple
times, the idea is to remove as much of the duplicate executions as
possible. At the same time, while merging to main we run duplicated
jobs, the idea is to remove that duplication.

**Description**

This PR accomplish a couple of things:
- It creates a new make tasks for running hive without depending on the
image build
- It generates a new job in the hive workflow to build the image once
and then share it with every simulation
- Merges the assertoor stability check with hive to leverage the docker
image already built
- Fix an issue were workflows were triggered two times in a row while
merging to main (both `merge_group` and `push` rules were set)

Resolves #1008
Resolves #1041