feat(l2): implement state diffs (#1015)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
With this PR, we now send the state diff (i.e. account updates, deposits
and withdrawals) via EIP-4844 blobs. This is cheaper than using EIP-1559
calldata.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
The Proposer now re-exceutes the block to get the account updates. With
this and deposits/withdrawals info, it generates a Blob encoding the
diff as mentioned in `crates/l2/docs/state_diffs.md` and send it in the
commitment step to L1.
For now, we'll use Ethereum's c-kzg-4844 library to do KZG operations.
The idea is to change it with Lambdaworks later.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #729