feat(levm): ethereum foundation tests setup (#1079)
> [!WARNING]
> DO NOT MERGE UNTIL #1029 IS MERGED.

**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

To have a working setup to run the EF tests to ensure the correct
implementation of our VM.

> [!NOTE]
> This setup only ensures that the pre-state is ok. We need further
development on the VM side to ensure the rest.

**Description**

- `crates/vm/levm/tests/ef/report.rs`. Implementation of
`EFTestsReport`. This is a data structure that keeps track of the tests
run.
- `crates/vm/levm/tests/ef/runner.rs`. Contains the implementation of
the EF Tests runner. It downloads the tests and executes each test as it
parses them. Test execution runs this flow: "prepare VM" -> "ensure pre
state" -> "execute VM" -> "ensure pos state". In this first iteration,
we only have the first two working as we need to continue working on
developing the VM to have them working.
- `crates/vm/levm/tests/ef/test.rs`. Contains all the types that
represent the objects from the EF tests `.json` files.
- `crates/vm/levm/tests/ef/deserialize.rs`. Contains the implementation
of the deserialization of the above.

**Try it yourself**

```
cargo test --release --manfest-dir crates/vm/levm/Cargo.toml testito
```

---------

Co-authored-by: JereSalo <jeresalo17@gmail.com>
Co-authored-by: Juani Medone <juan.medone@lambdaclass.com>
Co-authored-by: maximopalopoli <mpalopoli@fi.uba.ar>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>