feat(levm): move `levm` EF tests to `cmd/ef_tests` dir (#1172)
**Motivation**

We need `ethereum_rust` as a dependency for running the EF tests,
specifically for the post-state validation. If we import it inside the
crate `levm`, we have a circular dependency as `ethereum_rust` also
imports `levm`.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

- Move `ethereum_rust` EF tests to `cmd/ef_tests/ethereum_rust`.
- Update `ethereum_rust` Makefile targets that run the EF tests
according to the previous change
- Move `levm` EF tests to `cmd/ef_tests/levm`.
- Update `ethereum_rust` Makefile targets that run the EF tests to skip
`levm` EF tests run (this will be done in the crate `levm`)
- Update `levm` Makefile targets that run the EF tests according to the
changes above.