feat(l1): blob additional validations (#1178)
**Motivation**

We need to validate blob bundles through their commitments and proofs.

**Description**

Tis PR moves the Blob bundles validation to the blobs_bundle module in
the common crate, it perform early return validations regarding the
length of the blobs, commitments and proofs and after that validate both
the versioned hashes in the transactions as well as the commitments and
proofs generated.

Apart from that a couple of convinience functions were added, specially
a way to generate a `BlobsBundle` from blobs (with both commitments and
proofs) as well as a generation of the versioned hash from a
`BlobsBundle`. Apart from that a useful function from l2 utils was moved
to this module to be able to generate blobs from bytes.

Resolves #1200

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>