feat(l1): verify range (#1160)
**Motivation**
When we receive a range of data (be it storage, accounts, nodes) from a
snap request response we also receive the edge proofs for the given
rate. In order to validate that response and integrate the data received
we must verify the proofs in relation to the given range
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Implement function `verify_range` which verifies a given range of trie
elements with the provided edge proofs and also returns a boolean
indicating if there is more state to be fetched after the last key.

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None, but it is vital to process the responses from other nodes
during snap-sync #184

---------

Co-authored-by: Esteban Dimitroff Hodi <esteban.dimitroff@lambdaclass.com>