fix(l2): continue if address is not present when creating ExecutionDB (#1207)
**Motivation**

We were having this problem:
```
WARN ethereum_rust_l2::proposer::prover_server: Failed to handle request: ProverServer failed to create inputs for the Prover: Execution DB error: Missing account 0xB9e9A3651E41eDB1AceE24f58d743616dc8351c8 info while trying to create ExecutionDB
```

It was caused by sending a transaction to an account that isn't present
in the genesis file.
When constructing the ExecutionDB we are using the state of the `block's
parent`, this state doesn't have the account stored.

**Description**

If we don't raise an error if the account is not present, we fix this
problem.