build(core): remove jobs from pr and add to merge group. (#1260)
**Motivation**
There are jobs that are quite unrelated to the files that have been
changed, but still could cause an issue in an edge case. Let's run these
tests on the merge queue instead of CI.

The main idea is that a job should be either in pull request or in merge
group. Jobs on pull requests should be the jobs that test code that was
changed and merge queues tests code that is probably unrelated but we
want a sanity check before merging

**Description**
- Moved several heavy tests from CI (on pull request) to merge group
- Various formatting and consistency changes.
- Removed "ci skipped" hack since we already include those jobs in the
merge queue