fix(l1): remove "inconsistent internal tree structure" panics (#1288)
**Motivation**

The `Trie` library assumes that all tries are well-formed: this means
that if some node contains the hash of a child, then that child is in
the tire's database. If this doesn't hold, then a `get()` call panics
with an `inconsistent internal tree structure` message.

For the L2 prover we're using "pruned tries" which only store the
relevant nodes for the set of values touched in some execution, breaking
the previous hypothesis. One option was to capture the panics on the get
calls, but this wasn't safe because `Trie` has internal mutability. The
path chosen is to remove the panics altogether and replace them with
errors.

**Description**

- replaces mentioned panics for errors