ci(levm): improve speed for tests (and CI) and solve depth problem (#1435)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- In some cases we run into stack overflow when executing ef tests. This
can be solved by executing in a more optimized mode.

- Add things that I think convenient to EFTests and CI.

- This will be helpful for CI in general and for [this other
PR](https://github.com/lambdaclass/ethrex/pull/1424) that I'm working
on.

**Description**
- Disables spinner by default -> This is because it improves performance
a lot
    - LEVM EF Tests Execution with spinner: 6 minutes
    - LEVM EF Tests Execution without spinner (with prints): 30 seconds
- Changes `Makefile` to use `release` mode for running tests
- Creates a `Makefile` for the folder in which the report is printed.
This was made in order to access the report file from the same directory
in which we ran `make run-evm-ef-tests`
- Adds verbose flag for printing tests names. This can be extended for
printing more things but before they were being printed by default and
we found that in a basic debugging scenario we don't usually look at
those prints.
- Enables adding flags to make run-evm-ef-tests that facilitate the
command for testing

Example of flags:
`make run-evm-ef-tests flags="--tests stEIP1559"`


<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

**Why use --release**
I tried to avoid using the --release profile and wanted to update our
test profile for running with optimizations by default but I had issues
doing it because LEVM is a crate of the repository, and:
[Cargo only looks at the profile settings in the Cargo.toml manifest at
the root of the workspace. Profile settings defined in dependencies will
be ignored.
](https://doc.rust-lang.org/cargo/reference/profiles.html)
I didn't find a simple way to override the test profile without
affecting the rest of the crates from the repository, so I decided it
wasn't worth the hassle because the benefits from it weren't as high.


Closes #issue_number