fix(levm): fix various kind of errors in opcodes (#1424)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Generic Call works well for basic call cases but it handles poorly
some scenarios, and we still have to figure out how to solve the stack
overflow problem when depth is too high.
- Fix other things I see along the way

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- It replaces `OpcodeSuccess::Stop` for `OpcodeSuccess::Continue` in
some cases because we don't want to stop execution of current callframe,
we just want to push to it's stack and continue.
- It fixes `op_gas`, that previously was calculated based on used gas in
environment, and it should be based upon used gas in current callframe.
- It fixes `op_gaslimit`, that previously returned the transaction
gaslimit, when instead it should return the block's gas limit.


**Status**
- Tests passing with current changes: 1930/4100
<!-- Link to issues: Resolves #111, Resolves #222 -->