fix(l1): fix hangs on the tx spammer (#1413)
**Motivation**

The spammer hanged in one of 2 forms, either it wait for tx that weren't
added in the second phase of each spam and timeout after 5 minutes
resuming the execution or it hangs ad infinitum if the issue happens in
the first phase. This indicate that we were losing transactions.

**Description**

This is a mid-term solution, the explanation about what happened is done
on the issue, [in this particular
comment](https://github.com/lambdaclass/ethrex/issues/1120#issuecomment-2520574797),
but TLDR: we were receiving `engine_getPayload` request with old
`payload_ids` and we filled *AGAIN* the block but with new transactions.
the block was discarded as orphan by consensus and we lost those txs
from mempool on the immediate subsequent request. This was caused
because we were missing the update of the payload when it was modified
in the second step of our build process.

The current solution stores the whole payload, i.e. the block, block
value and the blobs bunde. Given our current implementation (a naive
2-step build process) we either filled the transaction and ended the
build process or not, a simple closed flag was added to the payload
store to signal this and avoid refilling transactions, this is clearer
than just check if there are any but could be changed if preferred.

| Spammer logs             |  Dora explorer |
:-------------------------:|:-------------------------:

![image](https://github.com/user-attachments/assets/f768879b-4bba-41f6-991b-e81abe0531f4)
|
![image](https://github.com/user-attachments/assets/bd642c92-4a99-42fa-b99d-bc9036b14fff)

**Next Steps**
Enhance the build process to make it async and interactive instead of
the naive 2-step all-or-nothing implementation we had right now.


<!-- Link to issues: Resolves #111, Resolves #222 -->

Resolves #1120