refactor(l2): l1_watcher make deposits unique (#1321)
**Motivation**

We need a way to identify deposit. In order to do it, we can use a
"depositId" that works as a counter. It is being incremented every time
a deposit event is emitted.

We've been starting the `l1_watcher` from block 0. Fetching logs from
block 0 will throw errors if asking light_nodes and it's wasteful. We
will need a more robust way to keep track of it, for now, we should
start from the first block in which the `CommonBridge` was deployed.

**Description**

- [CommonBridge] Use the `depositId` to construct the deposit_hash
(`keccak(to || value || depositId)`)
- [L1Watcher] Use the `depositId` from the logs to avoid duplicating the
mint transaction. And use it as the mint transaction's `nonce`. In this
way, the [Committer] can create the `StateDiff` to send the `commit`
transaction to the OnChainProposer.
- Add useful Makefile targets.
- Add simple command to fetch the `depositId`.
- Start from the block in which the `CommonBridge` was deployed.

Closes #1191