fix(levm): avoid jumpdest in push args (#1434)
**Motivation**

The motivation is to do not count the push arguments in the valid jump
destinations.

**Description**

This is done by calculating the valid jump destinations when the vm is
initialized. This PR also adds the valid_jump_destinations attribute to
callframe, and uses it to determine valid jumps in is_valid_jump_addr().
This PR has its origin in making the vmIOandFlowOperations tests pass,
and also is inspired in [execution-specs
implementation](https://github.com/ethereum/execution-specs/) about
valid jump destinations.

Note: These changes caused me problems with the benchmarks, since they
created the vm with an empty bytecode and then set a new callframe
outside of the new one. What I ended up doing is that the init of the
callframe in which a bytecode is passed also calculates the valid jump
dests, which makes the benchmarks go a bit slower.

---------

Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>