fix(levm): fixing selfdestruct (missing check, destroy accounts at the end of transaction) and modifying substate (#1448)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

- We weren't checking if the account that executed SELFDESTRUCT had
positive balance, that is useful for gas calculations
- Created set of `created_accounts` and use it in selfdestruct to know
if an address was created in that transaction. -> Add to that set the
address in `Create` transactions and opcodes.
- "Delete" (remove from cache) the accounts registered to be destroyed.
- In Cancun it is enough to remove them from cache because the accounts
that are registered to be destroyed are always accounts created in the
same transaction, so they never existed in the database in the first
place. But we will have to change this when thinking of implementing
this for previous forks.
- I moved touched accounts, touched storage slots and the selfdestruct
set to the substate, so that it reverts on execution


Tests passing: From 1941 in main to 1989 in this branch. Every little
change in this PR made some tests pass so there is no change that breaks
things.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1441