fix(levm): initialization of transient storage (#1540)
**Motivation**

Each time we moved to a new context, we were overwriting the previous
transient storage. Also add a check in `tstore` to see if the `tx` is
static.

**Description**

- Add a new parameter to `CallFrame::new()`. Now we can send the current
value of `transient_storage` to the subcontext.
- Check in `tstore` if the `tx` is static. In that case it must returns
an error.

We can also think about moving the `TransientStorage` from `CallFrame`
to the `Environment`, since it is a storage of the tx.

**Reference**
[EIP-1153](https://eips.ethereum.org/EIPS/eip-1153): _If the TSTORE
opcode is called within the context of a STATICCALL, it will result in
an exception instead of performing the modification. TLOAD is allowed
within the context of a STATICCALL._