fix(levm): resize memory in `generic_call` before any validation (#1556)
**Motivation**

The `CALLCODE` operand was not resizing the memory with its new size; it
was only using the new size to calculate the gas cost. This turned out
to be a problem common to all the `CALL*` family of operands.

**Description**

Right at the beginning of `generic_call`, we now resize the memory of
the current call frame. This is because, the memory should be expanded
even if the caller does not have enough funds to proceed with the
transaction.

This comes from the test `stRandom/randomStatetest248.json`. In this
test, the caller is trying to transfer `U256::MAX` to a different
address. In our code, we only expand the memory IF the user has enough
funds; but we should expand the memory regardless of funds.