feat(l1, l2, levm): `make stats` (#1361)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

We want to be able to run all the repo statistics locally with one
command (`make stats`).

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

- Updates `loc` bin (lines of code reporter).
- Adds `stats` target to the Makefile that runs:
    - Lines of code report
    - LEVM EF Tests
    - Hive simulation

**Work in progress**

Right now, the hive reporter needs some improvements. If you run the
command `make stats, ` it will run all and print the reports for lines
of code, LEVM EF Tests, and Hive, but the Hive report is off.

**Try it yourself**

```
make stats
```

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>