feat(l2): add encode_calldata function to abstract ABI encoding (#1593)
**Motivation**

This PR adds a proper `encode_calldata` function that encodes calldata
generation for any solidity function. To support this, a new `Value`
struct has been added, representing the possible solidity types for
arguments:

- `Uint` -> `uint256`
- `Address` -> `address`
- `Bool` -> `bool`
- `Bytes` -> `bytes`
- `String` -> `string`
- `Array` -> `T[]`
- `Tuple` -> `(X_1, ..., X_k)`
- `FixedArray` -> `T[k]`
- `FixedBytes` -> `bytesN`

This PR also moves some of the modules present in the `l2/utils`
directory to the l2 sdk instead.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>