fix(l1): block gas limit (#1597)
**Motivation**

The `block_gas_limit` is around 30M gas per block. The calculations
automatically adjust the gas limit if it changes. When using the `min`
function, the gas limit is consistently decremented. Consequently, if
the node runs for an extended period, it may eventually reach a point
where no transactions are accepted.

**Description**

Change `min` with `max`. With this change, the gas reaches 30M and stays
around it.