feat(l2): multiple provers in file (#1477)
**Motivation**

To support multiple proving systems, we may want to store the proofs and
wait to have all the needed information to send the transaction to the
L1 and verify them on chain.

In the previous PR (#1466) a HashMap was proposed to store the proofs in
memory, a more robust solution is needed in case the L2 node goes down.
In order to avoid re-execution, the proofs can be saved in files.

**Description**

- Add a module that stores the proofs in files.
- Instead of saving the proofs in a HashMap save them in files with the
help of the module.

- This PR doesn't delete old proofs if the `prover_server` throws an
error when resending the `verify` transaction, in that case, the flow
will be interrupted, and when the process starts again it will send(or
wait for the proofs) the verify transaction with the new block_number.

Depends on #1466
Closes #1060