fix(l2): sdk deploy (#1619)
**Motivation**

We were having some issues related to the `deploy()` function
implemented in the sdk. It was a problem derived from `TxKind::Create`

**Description**

- When doing gas_estimations we were setting the `to` field of the json
request to `0x0` if it was previously set to `TxKind::Create`. Now, we
set it to an empty String. In this way, the node deserializes it to
`TxKind::Create` by default.
- We were calculating the contract's address incorrectly, using the same
approach as the `levm` crate.
- Change the fields inside the `Overrides` struct to make it more
compatible with `EIP1559`:
  - gas_price &rarr; max_fee_per_gas
  - priority_gas_price &rarr; max_priority_fee_per_gas
- Update the `thiserror` crate and reference the workspace's version
when possible