fix(l1): fix the transactions spammer count diff with our node (#1650)
**Motivation**

As explained in [this
comment](https://github.com/lambdaclass/ethrex/issues/1449#issuecomment-2531681111)
the spammer has a 2-phase execution, where `txcount * accounts` define
the number of transactions added in the second phase of each round. We
were recieving significant less transactions in the secod phase, near
`accounts` in number

**Description**

This PR solves the issue adding a implementation of
`eth_maxPriorityFeePerGas`, which after debuging proved to be the cause
of the diminished number of transactions reaching our node. After an
initial implementation we can see the node working as expected, here is
a comparision:

| Before `eth_maxPriorityFeePerGas` | After `eth_maxPriorityFeePerGas` |
| ------------- | ------------- |
|
![image](https://github.com/user-attachments/assets/445516b9-3dc7-4b52-b78b-d314ac2d50d6)
|
![image](https://github.com/user-attachments/assets/54ba9bd9-14eb-4076-81d1-df5328a1ca89)
|

For the current implementation we extracted the common logic
between`eth_GasPrice` and `eth_maxPriorityFeePerGas`. This was inspired
in how geth managed both. I moved the logic from `gas_price` to a common
module (`fee_calculator`) that calculates the gas tip; previously we
were estimating the gas price, now the tip, this was the only change to
the logic. Now both RPC calls depend on this logic, the difference being
`gas_price` adding the `base_fee` while `max_priority_fee` doesn't. I
tried to avoid as much changes as possible to the previous
implementation of `gas_price`, this is something that we might want to
look further in the future.

Resolves #1449