feat(l1): add `show_peer_stats` method (#1659)
**Motivation**
When running the node it is useful to periodically see how many peers we
have registered and how many we are currently connected to. This PR adds
a method that recounts how many peers we have, how many of them are
active, and how many support the snap capability. In the future we can
extend this to show active peers supporting other capabilities too.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add a method `show_peer_stats` to the kademlia table which displays
how many snap/active/total peers we currently have
<!-- A clear and concise general description of the changes this PR
introduces -->

**Follow-Up Work**
I couldn't find a suitable loop where to add this method call so it is
only a debug utility for the time being until we find a suitable place
or decide to add a periodic task for it.
<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None