ci(l2): automate flamegraph tests (#1366)
**Motivation**

We want to generate flamegraphs both for `ethrex` and for `reth` on each
push to main, and then be able to view them in github pages or similar.

You can see the flamegraphs at https://lambdaclass.github.io/ethrex/

**Description**

- Github workflow for generating the flamegraphs and deploying the .svg
files to Github Pages
- The `ethrex_l2`CLI was modified to be able to create a default config
non-interactively with `ethrex_l2 config create default --default` since
the interactive CLI was not working in Github CI.
- `ethrex_l2 test load` was modified so that it now retries when it
can't connect to the server instead of failing directly, this was needed
to run the test in Github CI.
- `test_data/genesis-load-test.json` was modified to add balance to a
test account, since the tests were failing sometimes otherwise.
- You can run `make flamegraph` locally to run the same scripts locally
to generate the Ethrex flamegraphs (not the Reth ones).

**Notes**
- I had to run the perf record and then `inferno-collapse-perf` plus
`inferno-flamegraph` separately because of a bug in `perf` with Github
CI where it would output garbage to the stdout instead of the actual
perf data to a `perf.data` file.

**Things that can be improved or we can check**

- Make sure the flamegraphs are being correctly generated, that the
symbols are resolved properly. We saw some weird things where some
symbols were duplicated.
- We are also calling `perf script` with the `--no-inline` option, since
it takes a very long time without it, maybe we're not getting the proper
data because of that.
- Check if we can cache Reth so we don't compile it at each time.
- Maybe the web can be improved, We can generate `.png` files as preview
since the `.svg` look cropped there.
- `scripts/flamegraph.sh` and `.github/scripts/flamegraph_watcher.sh`
and very similar and they could be merged into one, however I hardcoded
some CI-paths in the latter one.

<!-- Link to issues: Resolves #111, Resolves #222 -->

---------

Co-authored-by: fborello-lambda <federicoborello@lambdaclass.com>
Co-authored-by: Federico Borello <156438142+fborello-lambda@users.noreply.github.com>