fix(l1): use levm flag for apply transaction function (#1755)
**Motivation**
We are using revm for the
[apply_plain_transaction](https://github.com/lambdaclass/ethrex/blob/213af945b71c9a61d1966054a6ee97fad4abc770/crates/blockchain/payload.rs#L421)
function; regardless of the EVM implementation passed as a CLI argument.

With this PR, said function will now use the EVM implementation
specified via CLI flags and will default to LEVM.

**Description**

Add both the LEVM and REVM code inside the apply_plain_transaction
function. Only the specified implementation gets compiled.

Closes #1746