feat(levm): add debug build directive for run-evm-ef-tests (#1780)
**Motivation**

If one needs to run LEVM with debug symbols (in order to use a debugger,
such as lldb), the `--release` flag in the makefile has to be removed
manually every time.

This PR adds a new if branch that checks if the "DEBUG" environment
variable was passed

**Description**

With this PR, running:
```
make DEBUG=true run-evm-ef-tests 
```
will compile the project with debug symbols. 


Closes #1779