fix(l2): replace `ExecutionDB` `from_exec()` impl to use a `CacheDB` based approach (#1709)
**Motivation**

A `ExecutionDB` is created from a block's pre-execution but the current
implementation is incorrect and not all needed data gets stored. The
goal is to replace this with another approach, using an auxiliary
"caching" database that will retrieve data from a `Store` or an RPC
endpoint (like in #1705) during the pre-execution of a block, finally
resulting in a db that contains all the needed execution data.

**Description**

- replace `ExecutionDB::from_exec()` to use the `CacheDB`, rename it to
`from_store()`
- some refactors