refactor(levm): clean `vm.rs` file (#1796)
**Motivation**

The `vm.rs` file is filled with functions that are used as "helper"
functions. It would be better, especially for newcomers to the project,
if the `vm.rs` file had only the bare essential "big picture" functions.


**Description**

This PR aims to move a lot of the "additional" logic to `utils.rs` file
and only keep the "big picture functions" in `vm.rs`

This PR moved the following functions from `vm.rs` to `utils.rs`:
- address_to_word
- calculate_create_address
- calculate_create2_address
- get_valid_jump_destinations
- get_account
- get_account_no_push_cache
- get_account_mut_vm<'vm>
- increase_account_balance
- decrease_account_balance
- update_account_bytecode
- get_intrinsic_gas
- max_blobs_per_block
- get_blob_base_fee_update_fraction_value
- get_base_fee_per_blob_gas
- get_max_blob_gas_price
- get_blob_gas_price
- get_n_value
- get_number_of_topics
- increment_account_nonce
- decrement_account_nonce
- word_to_address
- has_delegation
- get_authorized_address
- eip7702_recover_address
- eip7702_get_code
 

And deleted the following: 
- VM::get_account
- VM::get_account_no_push_cache
These were simple wrapping functions around functions that had the same
name.

The logic inside the functions was left unchanged, besides adding
parameters to the functions.
Closes #1632