perf(levm): optimize op_shl (#1841)
**Motivation**

Reviewing the benchmarks for the ERC20 contracts, a major bottleneck was
the constant bitshifting we were making, since `checked_shift_left` is
quite slow. This introduces certain precomputed values and small
optimizations to make that faster.

**Description**

* We have a table of a few common precomputed bitshift values for
`1<<amount` (units like wei, szabo, eth, etc. , addresses).
* If the value is not precomputed, we check if it's a power of two
(`1<<amount`) and do a safe pow() operation to avoid calling
`checked_shift_left`.
* Else it runs normally like it used to.

**With this we go from a 12-10x difference with revm in the ERC20
benchmarks to only a ~2x more in par with other benchmarks**. There's an
obvious catch to this and it's that it's because we precomputed
`1<<160`, but this is used for calculating addresses commonly in
Ethereum contracts, so it should be useful imho.

I'm obviously open to any discussions since this does add a bit of
complexity and it's a tradeoff, but the code is pretty simple.

---------

Co-authored-by: fborello-lambda <federicoborello@lambdaclass.com>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>