feat(core): add eip7702 tx (#1804)
**Motivation**

We have to add the [EIP-7702: Set EOA account
code](https://eips.ethereum.org/EIPS/eip-7702) transaction to our
implementation.

**Description**

- Add de/serialization
  - Simple de/serialization test
- RLP de/encoding
- Common methods for `Transaction`/`GenericTransaction`

>[!IMPORTANT]
> The integration with L1 was achieved by updating `revm` to version
`18.0.0`, which is the latest version compatible with our setup. This is
because we can only use `revm-inspectors` up to version `0.13.0`, as
later versions require Rust `1.82.0`, which would break the prover.
>
> LEVM should be up-to-date and functional while running EIP7702
transactions, but REVM will be outdated.

Closes #1792

---------

Co-authored-by: avilagaston9 <gavila@fi.uba.ar>
Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>