feat(levm): add hooks support and implement `DefaultHook`  (#1869)
**Motivation**

LEVM neeeds to support a L2 feature called "Privilege Transactions". To
achieve this, some functionality of the EVM has to be able to be
customized. In order to achieve this, the EVM will start to support
"Hooks" which will modify the `prepare_execution` and
`finalize_execution` functions of the EVM.

**Description**
This PR implements the `DefaultHook`, which is a hook that maintains
backwards compatibility with LEVM's previous way of operation; whilst
allowing the introduction of future alternative hooks.

In order to achieve this, the following changes were made:

- [X] Move `eip7702_set_access_code` to a standalone function.
     - [X] Move `access_account` to an standalone function.
- [X] Move `add_intrinsic_gas` to a standalone function .
- [X] Make `beacon_root_contract_call_levm` receive a `Arc<dyn
LevmDatabase>` instead of a `Arc<StoreWrapper>` to make it more like
`execute_tx_levm`.
- [x] Replace `&Arch` with simply `Arc` in `utils.rs`.
- [x] Implement `DefaultHook::prepare_execution`.
- [x] Implement `DefaultHook::finalize_execution`.

Closes #1879
Closes #1880
Closes #1629

---------

Co-authored-by: Francisco Krause <francisco.krause@lambdaclass.com>