chore(levm): rename EF tests from `levm` to `state`. (#1888)
**Motivation**
To be more aligned with the nomenclature used by the EF.


**Description**
We have "blockchain tests" and "state tests". These ones are state
tests, see:
https://ethereum.github.io/execution-spec-tests/main/consuming_tests/state_test/

Adding `levm` is confusing since we could also run them with revm. But
instead of calling them evm tests, let's stick with state so it is
aligned with how Ethereum calls them.