refactor(levm): clean out address_access_cost (#1935)
**Motivation**
This PR is a continuation of #1932.
The function `address_access_cost` contained an if statement that simply
returned a static cost value when a specified fork was active.

However, sometimes the opcodes that called the function contained other
if statements that overlapped with the ones inside
`address_access_cost`.

**Description**

- Remove the if statements from the `address_access_cost` and moves that
logic to the opcodes.

This PR Also fixes some EF-Tests as a consequence.