refactor(l1, l2): move rpc clients to `rpc/clients` module temporarily  (#1984)
These clients were wrongly distributed in the codebase as they were only
used by L2 crates or `blockchain`'s `dev` crate. For preconfirmations,
we need to access these types at the `rpc` crate level as we need to
intervene in some of the RPC API endpoint handlers to relay the request
to an external gateway; and importing them from where they are today
generates a cycle import.

This is not the right place for them (`rpc/clients`), as we should
expose an SDK separate library, but this means a more complex refactor.
This was the easiest change to fix the cycle dep problem.