feat(l1): add `new_from_genesis` method to `Store` (#2007)
**Motivation**

Without this method, users must do the genesis file decoding manually,
sometimes needing to unnecessarily add the `serde_json` crate to their
`Cargo.toml`.

Having a method that initializes the genesis state in the `Store` allows
the user to start a new `Store` from the genesis file, which only has
the genesis path.

**Description**

Adds a new `new_from_genesis` method to `Store` that initializes a new
`Store` instance with the genesis (provided as a path to the genesis
file).