fix(levm): increase support for `stZeroCallsTest` test suite (#1975)
**Motivation**

This pull request includes several changes to improve handling of
account state transitions and gas cost calculations, particularly in
relation to Ethereum's EIP-161.

**Description**

* `cmd/ef_tests/state/runner/levm_runner.rs`: Implemented logic to
remove accounts that remain empty after the Spurious Dragon fork, in
compliance with EIP-161.
* `crates/vm/levm/src/gas_cost.rs`: Added a new parameter
`address_exists` to the `call` function and updated the logic for
handling empty accounts based on EIP-161.
* `crates/vm/levm/src/opcode_handlers/system.rs`: Updated the `call` and
`selfdestruct` function to check if an account exists and adjusted the
logic for determining if an account is empty based on the fork
configuration.

---------

Co-authored-by: Tomas Fabrizio Orsi <tomas.orsi@lambdaclass.com>