feat(l2): Switch to a system contract to handle user withdrawals (#2012)
**Motivation**

This PR introduces a `CommonBridgeL2` system contract (that comes with
the L2 genesis) at address `0x000000000000000000000000000000000000FFFF`
that handles withdrawals. To issue a withdraw, users simply send a
transaction to the `withdraw` function with the address of the account
that will claim the funds on L1.

This was done because the previous approach to withdrawals was extremely
hacky and error prone, issuing a transaction to the L1 address that
would claim, that was then internally changed to the zero address before
vm execution. This approach is a lot simpler and does not require any
hackyness of that kind.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number