chore(l1): move `ethrex-trie` to common folder. (#2051)
**Motivation**
Even though the trie is related to a structure that is stored on disk,
we have decoupled the trie logic from the storing (see `trie_db`).

Some modules like the types, just use the trie logic to calculate hash
tree roots, without storing anything, for example:
https://github.com/lambdaclass/ethrex/blob/51b7dcbac7226e748b711ced1920bcf11b326577/crates/common/types/account.rs#L158

Geth does something similar with:
- https://github.com/ethereum/go-ethereum/tree/master/trie
- https://github.com/ethereum/go-ethereum/tree/master/triedb

So the trie logic should go in `common`, whereas the trie db
implementations should go in `storage`.

Finally, this would simplify the dependencies if we decided to merge
common, trie, and rpc into one crate.