fix(l1, l2): handle invalid RPC requests (#2086)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
When handling RPC requests, if the body cannot be parsed to a valid
`RpcRequest`, the thread panics. Although it's an independent thread per
request, it might be better to successfully return an error response to
the client instead of panic and suddenly close the connection.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
If the body cannot be parsed to a valid `RpcRequest`, the following
message is sent to client:

```
{"id":"","jsonrpc":"2.0","error":{"code":-32000,"message":"Invalid params: Invalid request body"}}
```

<!-- Link to issues: Resolves #111, Resolves #222 -->