chore(core): assertoor check proposal validate txs published by every client (#2101)
**Motivation**

We want to check we don't have regressions after fixing pooled
transactions in #1990

**Description**

Now that we share txs with the other nodes we want our assertoor el
stability check to validate EVERY client and check everyone is able to
publish a block with the total 240 txs from the spammer.

This PR also re-enable the blob transaction check after ~10~ 15+ runs
without issues. If some run start to fail again as before randomly we
could disable it once more.