perf(l2): don't recompute transaction senders when building blocks (#2097)
**Motivation**

This PR modifies the code so that, when building payloads, we don't
recompute the transactions' senders, but rather use the one that was
already calculated when added to the mempool.

This represents a roughly 10-15% performance improvement on the
`build_payload` function. For L1 this is mostly irrelevant, as it will
almost never build payloads itself; for L2 this matters, as the L2 has
to build (sequence) blocks itself.

Requires merging #2090 first

Flamegraph recomputing the address:

<img width="1506" alt="without_caching"
src="https://github.com/user-attachments/assets/938723b6-a62c-4cf9-8d98-d1fd50b2d145"
/>

Flamegraph using the cached value:

<img width="1511" alt="with_caching"
src="https://github.com/user-attachments/assets/7f421460-b6e3-42bd-b2d1-b2a40d967d38"
/>

Note that in the first flamegraph the `execute_tx` function spends a
decent part in the `tx_env` function, which is where the sender's
address is computed. On the second one `tx_env` is gone.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Estéfano Bargas <estefano.bargas@fing.edu.uy>