fix(l1): receipts invalid size libmdbx (#2065)
**Motivation**
Syncing holesky.

**Description**
While syncing on Holesky, we encountered blocks that exceeded the
maximum allowed entry size in libmdbx. Increasing the page size was not
a viable solution, as even the maximum allowable size in libmdbx wasn't
enough for some blocks.

To address this, we implemented an `IndexedChunk` trait that splits
oversized data into smaller chunks, ensuring each piece stays within the
maximum value size limit. Since libmdbx is structured as a B+ tree map,
we had to ensure correct order of chunks. We achieved this by appending
a byte representing the chunk index, this way we preserve the correct
sequence of chunks during retrieval.

Moving forward, we should consider whether other tables might also
exceed this size limit. If so, we can extend this trait.

Closes None