fix(l2): check if `AccountUpdate`'s `AccountInfo` is `None` in l1 committer. (#2133)
**Motivation**

When running the L2 load test (`make load-test-io`), you would get the
following error in the L2: `ERROR ethrex_l2::proposer::l1_committer: L1
Committer Error: Committer failed retrieve data from storage`.

This PR fixes this error in the L2's committer. 

**Description**

In the `l1_committer.rs` file, `AccountStateDiff`s were constructed from
`AccountUpdates`. More specifically, it used the `AccountUpdates`'s
`nonce`.
The problem was that `AccountUpdates` sometimes didn't contain `nonce`
as a value, and the whole `AccountInfo` struct (which holds the `nonce`)
evaluated to `None`.

Therefore, if the computed `AccountUpdates`'s had a `None` for
`AccountInfo`, it means that there was no change in `nonce`. Therefore,
the new nonce is equal to the nonce of the previous block.



<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2093

---------

Signed-off-by: Tomas Fabrizio Orsi <tomas.orsi@lambdaclass.com>