fix(l1): reorg syncing from sidechain (#2102)
**Motivation**

We still had 16 hive test failing related to syncing + reorg from the 31
available, this PR tries to find and address the issues

**Description**

The issues were related to a couple of reasons, here are them explained
and linked to issues for going further in depth:

- **Side chain Syncing:** We were unable to sync with side chains due to
how we requested headers (from the current head to the sync head), a
quick fix is implemented in this PR and Issue #2126 was created to
tackle it properly.
- **Invalid Ancestors management:** We were marking just the failing
ancestor in the sync cycle, we've implemented a quick fix in this PR to
also mark the sync head as invalid. This might be enough for the current
tests but something to take a look later in depth, specially related to
mark the entire chain as invalid or another alternative, the issue #2127
was created to follow that. Also another was was created to move in the
future the invalid ancestors to storage: #2136
- **Checking for an invalid block or parent**: This issue was two fold,
on one hand Fork choice updated lacked a check to make sure the parent
wasn't already marked as invalid, on the other NewPayloadV3 lacked both
check for the current block as well as the parent.

With these additions the 16 hive test failing related to Syncing + reorg
are now working.

---------

Co-authored-by: Marcos Nicolau <76252340+MarcosNicolau@users.noreply.github.com>