feat(l1): set max number of peers (#2026)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- Set max number of peers so that we stop connecting once we reach that
limit.

**Description**
- Max number of peers connected is calculated by looking at those peers
in the kademlia table that have `is_connected` set to true.
- When receiving a pong, don't initiate a connection if we have reached
max number of peers connected
- After handshake we start a connection, but before exchanging hello
messages I added post handshake checks. For now there is only one check
that sees if the max amount of peers has been reached. If that happens
we send a disconnect message and we finish the connection, removing the
peer from the kademlia table.
- I modified `connection_failed` method because I think we were sending
a Disconnect message even when we received a disconnect message from our
peer and that is not necessary. In those kind of scenarios we should
just disconnect I think
- It also creates the `DisconnectReason` enum so that we don't use
strings as reason anymore.

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1928

---------

Co-authored-by: Marcos Nicolau <76252340+MarcosNicolau@users.noreply.github.com>
Co-authored-by: Leandro Serra <leandro.serra@lambdaclass.com>
Co-authored-by: LeanSerra <46695152+LeanSerra@users.noreply.github.com>
Co-authored-by: Rodrigo Oliveri <rodrigooliveri10@gmail.com>