fix(l1): solve network mismatch when adding preset boot nodes (#2189)
**Motivation**
After a recent refactor the node was failing to add the preset boot
nodes for known networks, rendering the node unable to connect to the
network when using a known network without extra boot nodes.
This was due to the code first changing the value of the network flag to
the full path to the preset network's genesis file before adding the
preset boot nodes, so while the `get_bootnodes` function expected
`holesky` as network value it was instead receiving the full genesis
path, causing a mismatch. This PR solves this issue by matching against
the full genesis paths, and also removes misleading info tracing (there
was an adding preset bootnodes message when the network flag was being
changed to the full genesis path)
<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number