fix(l1): add prague timestamps to holesky & sepolia genesis (#2215)
**Motivation**
Holesky and Sepolia testnets have moved on to Prague but we haven't
registered this in their preset chain config, causing us to reject all
newPayloadV4 requests as we asume the block to be Cancun instead of
Prague. This PR fixes this by adding the Prague timestamps for both
networks.
The timestamps were taken from
[geth](https://github.com/ethereum/go-ethereum/blob/f3e4866073d4650a7f461315c517333c6407ab5c/params/config.go#L99)