chore(l1): revert conditional docker building logic (#2196)
**Motivation**

#2175 introduced an additional compilation check regarding L1 client for
the hive tests. The check was that to avoid building the `ethrex` docker
image if it was not being used.

The check added additional complexity whilst not providing a lot of
utility, since the ethrex docker image would have to be built
regardless; since the only point of using a different L1 Client was to
compare against `ethrex`.

**Description**

Remove the if statement that provided the conditional compilation of the
`ethrex` docker image.

Also remove an additional `HIVE_LOGLEVEL` that was not present in the
Makefile before.

---------

Signed-off-by: Tomas Fabrizio Orsi <tomas.orsi@lambdaclass.com>