fix(levm, l1): fix hive tests rpc (#2176)
**Motivation**

- When executing ethereum/rpc simulation with LEVM we have errors in a
lot of tests. The error is when executing the blocks for the setup for
these tests and how we created the withdrawals requests introduced in
EIP-7002

**Description**
- First, when we execute the transactions, we weren't handling if the
account was delegated and we weren't using the correct bytecode.
For this, we add the check for delegated accounts and get the bytecode.
This is made when creating a new VM and using the function
`eip7702_get_code()`.
- Second, when creating the withdrawals requests, we were overwritting
and loosing some of the new values in the account updates. We now merge
both updates.
- When running the system contract for the withdrawals requests
introduced in EIP 7002, we were creating a new `CacheDB` instead of
using the one with updated values.

**Status**

- All RPC Hive tests were passing, but with the changed made
[here](https://github.com/lambdaclass/ethrex/pull/2179), 12 are failing
due to a new missing implementation.

Links to #2158

---------

Co-authored-by: Tomás Paradelo <tomas.paradelo@lambdaclass.com>