feat(core): add `p2p.enabled` flag (#2230)
**Motivation**

In most of the L2 use cases we want to disable de P2P network.

**Description**

Add a `p2p.enabled` flag for users to explicit whether they want to
enable the P2P in their node.

It is enabled by default in the L1 and disabled by default for the L2.