fix(l1): fix unending storage healer process in snap sync (#2287)
**Motivation**
There is currently a bug in snap sync. When a state sync becomes stale,
the snap sync cycle is aborted but the storage healer process is left
hanging instead if signaling it to end and waiting for it to finish. The
loop condition of the storage healer is also not properly set, keeping
it alive even after the end signal if it still has paths to heal. This
PR fixes both of this problems
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Fix loop condition in storage healer
* End storage healer if state sync aborts due to stale pivot
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number