feat(l2): add state reconstruction command (#2204)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We need a way to reconstruct the chain state in case of a failure or
even if someone want to _trustlessly_ access the state.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
This PR introduces a new ethrex_l2 CLI command, `stack reconstruct`,
that takes downloaded blobs from L1 and reconstruct the blocks based on
its info, storing the state in a Libmdbx store. The blobs can be
downloaded using the `stack blobs-saver` command.
At this stage, the command is able to successfully reconstruct the chain
state and continue to produce blocks.
Note that, as we send state diffs and not transactions to L1, some data
(i.e., transactions history, receipts) will not be accessible in a
reconstructed network.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1103

---------

Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>