refactor(l2): separated configs for prover_client and sequencer (#2269)
**Motivation**

When running the prover_client as a standalone component the
`config.toml` wasn't being parsed.
Ideally we should parse it before we run the prover_client.

**Description**

- The `ConfigMode` enum is proposed to parse the .toml for the
`Sequencer` or the `ProverClient`
- The prover_client parses the `prover_client_config.toml` and creates a
`.env.prover` file
- Created new envars to set the:
  - `CONFIGS_PATH` 
  - `SEQUENCER_CONFIG_FILE`
  - `PROVER_CLIENT_CONFIG_FILE`
  - The references were updated in the Makefile

This change also enables us to change the `SEQUENCER_CONFIG_FILE` easily
keeping it in the `configs` dir and switching the `Makefile`'s variable.
(Useful when testing locally and with a testnet).

Closes #2053

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>