feat (l1): write nodes in batches during storage healing (#2288)
**Motivation**
In a similar fashion to #2270, this PR aims to reduce the time spent
writing data to the DB by writing data in batches. In this case the
nodes received during storage healing are written all at once using the
already existing `put_batch` method of the TrieDB trait.
This could only be done for nodes belonging to the same trie, as it
would otherwise involve leaking and constraining the internal
representation of TrieDB.
This has shown to reduce the time spent writing storage nodes in the DB
from around 4 seconds to less than 20 milliseconds
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add `write_node_batch` method for `TrieState` relying on
`TrieDB::put_batch`
* Refactor storage healer code to write all nodes for a trie in a single
operation
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Rodrigo Oliveri <rodrigooliveri10@gmail.com>