perf(levm): remove repeated `get_account_info` calls in LEVM (#2357)
**Motivation**

Noticed on #2292 that the majority of the time in
`LEVM::get_state_transitions()` was spent on calls to
`get_account_info()`. While looking for areas to improve, I found that
we were calling `get_account_info()` three times instead of reusing the
value returned in the first call.

**Description**

Removes the repeated calls to `get_account_info`.

Testing locally shows a `2x` speed improvement in
`payload_builder::build_payload()` implemented in #2292.

Closes None