fix(levm): fix hive tests regression levm (#2386)
**Motivation**
- Fix hive tests that were failing (tested all RPC tests and they pass)
<!-- Why does this pull request exist? What are its goals? -->

**Description**
- Created `stateless_execute()`, a wrapper of execute() that doesn't
alter the state. It actually backs up the cache before executing and
then restores it. I know cloning the cache is not ideal but
optimizations to that can be performed later. We use this method when
creating access list.
- Clear the cache when instantiating a new VM: our storage slots have
"original value" and "current value", this is for gas costs, but when
executing a new transaction it should happen that all the original
values and current values have to be the same. Before the database
refactor we were doing this outside of LEVM, now it's embedded in the
code!

Update: I'm currently trying to fix the failing test from
`ethereum/engine` suite. It is the test with name "Replace Blob
Transactions (Cancun) (ethrex)"

Closes #issue_number