refactor(levm): maintain cached state between executions in levm (#2371)
**Motivation**

This PR refactors levm introducing a `GeneralizedDatabase` (the name is
not the best, suggestions are welcome), which is simply the combination
of the regular store (i.e. the on-disk database) and the cache (the in
memory structure we use to keep track of changes done to accounts as
execution happens).

More importantly, it makes levm (the `VM` struct) hold a mutable
reference to said database instead of a regular owned value. This is to
bring the levm api more in line to what we need it to be for ethrex; the
main problem we are solving is to be able to run multiple transactions
while keeping the underlying cache changes from those transactions. Once
we are done with execution (because we've finished executing or building
the block or group of blocks), calling `get_state_transitions` will
drain the cachedb for all the account updates and return them.

This change allowed, among other things, to implement for levm the
previously uninplemented function
`execute_block_without_clearing_state`, which is used for syncing.

Additionally, this PR introduces error handling for the levm `Database`
trait; previously its member functions did not return errors and we were
just unwrapping on some of their implementations.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>
Co-authored-by: JereSalo <jeresalo17@gmail.com>
Co-authored-by: Martin Paulucci <martin.c.paulucci@gmail.com>
Co-authored-by: fmoletta <99273364+fmoletta@users.noreply.github.com>