fix(l1): move on to the next retry upon failed sends on peer handler requests (#2369)
**Motivation**
The PeerHandler contains methods to request all sorts of data from peers
and ins used both in snap and full sync. It uses a retry system to
ensure that we don't misinterpret a malicious/unsynced peer as the data
not being available. If one peer doesn't return the requested data we
would try with other peers first before assuming that the data is too
old or doesn't exist.
When we sent the requests to the peer, we were not respecting the retry
policy and returning a None response upon first failure. This led the
synced to believe that data had become stale when it was not the case.
This, on multiple occasions, caused the storage fetcher to cease
fetching while the state sync was still alive and fetching accounts,
leaving a huge backlog of storages to heal after the state sync.
This PR solves this by moving on to the next retry upon a send error
instead of aborting the request and returning an empty response
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Continue to the next retry instead of returning None upon a failed
`send` in all `PeerHandler` request methods
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number