fix(l1): add new EF tests which were previously being ignored (#2425)
**Motivation**

We have a few EF test which are currently being ignored, since they were
not passing.
Some of them are now passing, and others still fail. We want to fix them

**Description**

This PR accomplishes two things:
1. Adds back some EF tests that were previously being ignored because
they were failing and now pass.
2. Fixes one EIP-7702 test by making a modification on the test runner.

The EF test `eip7702_set_code_tx/set_code_txs/contract_create` validates
that an error is thrown if an EIP-7702 tx has the `to` field set to
null, as it is stated on the
[EIP](https://eips.ethereum.org/EIPS/eip-7702). The issue with this test
is that it expects for a `TransactionException` to be thrown, which in
our code may only happen if the `Transaction` data type is successfully
decoded from RLP.

Our code is effectively throwing an error, but that's happening while
RLP decoding the block which contains said transaction.
Since this wasn't being expected from the test, it was failing.

This PR includes a patch for this test, by including the
`TransactionException.TYPE_4_TX_CONTRACT_CREATION` to signal an expected
RLP decode error on the EF test runner.

Another solution was considered, which would have had to modify the `to`
field of the `struct EIP7702Transaction` to be a `TxKind` instead of an
`Address`, allowing for a `TxKind::Create` type to be decoded from RLP,
just to throw an exception later in the code, most probably on
`validate_transaction` or `prepare_execution` hook. This solution was
discarded in favor of the previously detailed one.

Something to note is that the type 3 transaction also has the same
behavior, and there is an EF vector test which validates it.
The test is `eip4844_blobs/blob_txs/invalid_blob_tx_contract_creation`
and it allows for both `BlockException` and `TransactionException`
errors to happen.

Most probably, the ethereum foundation forgot to add the
`BlockException` to the EIP-7702 test.