feat(l2): add prover client to integration test (#2291)
**Motivation**

The integration test is being run without a prover client and bypassing
verification. We want this test to run every component of the L2.

**Description**

- executes the prover client together with the integration test (ideally
we would run the prover in a docker container but some backends like SP1
start a container to execute the final phase of proving, a container
can't create another one without being privileged, so for now we run the
prover client as another process).
- refactors the state reconstruction test out of the integration test
- defines a workflow to run the integration test on each PR push using
the `exec` backend
- defines a workflow to run the integration test on main using the `sp1`
backend
- adds docker layer caching using `bake`, defines new "Docker bake" job
- removes any reference to zkvm-specific env variables used for
bypassing verification or mock proving
- removes the prover server dev mode loop. Now "dev mode" refers to
using the `exec` backend. This and the previous point make sure there's
always a prover client running and executing blocks, even if it's not
generating proofs.

Closes #2224

---------

Co-authored-by: fborello-lambda <federicoborello@lambdaclass.com>
Co-authored-by: Federico Borello <156438142+fborello-lambda@users.noreply.github.com>
Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>