feat(l1, l2): make some store getters async (#2430)
**Motivation**

Like with #2336 the goal is to avoid blocking the current task.

**Description**

Makes store getters not related to tries (and thus the EVM) async, and
propagates the changes to users of store. They are made async by using
`spawn_blocking `

Many instances of functional code (`and_then`, `map`) had to be replaced
due to bad async support.

Closes #2424