feat(l1): show healing in progress message at set intervals (#2462)
**Motivation**
We are not able to perform estimations on when healing will end, but we
should also not stay completely silent while healing takes place, as
this is not very user-friendly.
This PR aims to add messages to inform wether state and storage healing
are taking place, at the same pace as we show state sync and rebuild
progress. For state sync, pending paths will be shown. These can give an
insight on the progress, as the amount of paths will continuously
increase as we progress through the trie, but will start gradually
decreasing as we near the end of healing. For storages it is a slightly
different story as we don't have the full number of pending storages
available for showing so we only show the storages currently in the
queue.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Periodically show amount of paths left during State Healing
* Periodically show storages in queue during Storage Healing
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->