fix(l1): blob transaction init for revm ef tests (#2588)
**Motivation**

Fixing REVM tests as specified in issue #2491.

**Description**

This pr changes the `blob_excess_gas_and_price` variable initialization
in the `prepare_revm_for_tx` function. Now instead of setting gas price
in 0 I use the `new` function associated to the `BlobExcessGasAndPrice`
struct that sets the gas price by itself with a particular function that
calculates it.

This change drops the failing amount of tests from 2009 to 829.

This pr solves part of the issue #2491.