fix(l2): bench crate (#2647)
The pipeline was failing because of:
1. some breaking changes (renamed imports) 
2. a previously existing bug that we didn't catch until #2516 (we were
encoding a leaf node without its leaf flag, so it would be decoded into
an extension leaf and produce a panic)
3. we were executing the prover in L2 mode because we depended on
ethrex-l2 (and in ethrex-prover, which depends on ethrex-l2), and that
crate enables the L2 feature of ethrex-vm by default.
- The solution was to make an `l2` feature for ethrex-l2, which isn't
ideal. A better (but bigger) solution is proposed in #2662