feat(levm): replace ambiguous error with proper validation error when obtaining effective gas price (#2667)
**Motivation**
While implementing a mapper for our ethrex & levm error types for the
consume-engine hive test I ran into a test that was returning the error
`Invalid Transaction: Invalid Transaction` which doesn't look useful at
all. The error comes up when we fail to compute the effective gas for a
transaction (aka the block's base_fee is higher than the transaction's
max fee), so I replaced it with the appropriate error
(TxValidationError::InsufficientMaxFeePerGas) which is also the one
expected by the test suite.

**Description**
* Replace ambiguous error used when calculating effective gas price
before tx execution with proper validation error.

Closes: None, but is needed to cleanly implement the error mapper needed
for #2474