fix(l2): download solc fixed version in Dockerfile (#2700)
**Motivation**

A new version of solidity was released a few hours ago
([`0.8.30`](https://github.com/ethereum/solidity/releases/latest)) and
the Dockerfile was written to always download the latest version while
our contracts Solidity version is fixed to `0.8.29`.

**Description**

Updates the L1 contract deployer Dockerfile to download a fixed version
of solc.