fix(l1): fixed proper parent when receiveing a NewPayloadRequest (#2690)
**Motivation**

This pr aims to fix a bug with the parent assigned in NewPayloadRequest
if the parent is valid. From the [Paris fork
documentation](https://github.com/ethereum/execution-apis/blob/main/src/engine/paris.md)
"latestValidHash: DATA|null, 32 Bytes - the hash of the most recent
valid block in the branch defined by payload and its ancestors"

**Description**

Removed storage get for the canonical latest valid ancestor and replaced
with the parent hash (if it's valid).

Fixes 27 tests in #1285 in "engine-cancun"