fix(l1): catch potential panic when decoding `NodeHash` (#2683)
**Motivation**
The method `NodeHash::from_slice` can panic if the slice is over 32
bytes. This could cause panics when decoding nodes as it is used without
checking the length beforehand. This PR adds a check and returns an
invalid length error before calling `from_slice`. It also mentions the
potential panic on the method's documentation & removes a misleading
`From<Vec<u8>>` implementation that would also panic under the same
condition.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Remove `From<Vec<u8>> for NodeHash` impl as it could cause panics
* Mention potential panic on `NodeHash::from_slice` doc
* Check rlp decoded data len to avoid panics when decoding `NodeHash`
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2649