chore(l1): fix contract deployment tests from EIP-7002 (#2630)
**Motivation**

On #2586 new tests were added and some of them failed on LEVM and REVM.

**Description**

8 new tests are now working and dont need to be skipped in each of the
VMs. The tests were failing because we were not checking if the bytecode
of the system contracts that the EIPs
([7002](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-7002.md#empty-code-failure)
and
[7251](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-7251.md))
define were empty or not. And also because the we were not handling the
case were the system calls revert, leading to an invalidate block.

Closes #2598

---------

Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>