Move transfer condition in generic_call()
When doing a generic call (CALL, CALLCODE, DELEGATECALL,
STATICCALL) we were first checking if it was a delegation
and if the bytecode was empty, and if this was true the
call would not continue. After that the transfer
of WEI to the account was done. When using precompiles
(which have empty code) we never reached the transfering stage,
so the change in the balance was never done and the
state root was not correct. We now first do the transfer
and the the checking of bytecode. This fixes 17 tests that were
being skipped on LEVM.