perf(levm): optimize how levm tracks storage modifications in case of reverts (#2699)
**Motivation**

This PR replaces the way we track storage modifications to contracts
when executing to handle reverts. Previously, when writing to an account
we were cloning its entire accumulated modified state, so in case we
needed to revert we could overwrite it back to its former values. These
were the lines of code:

```
        let previous_account = cache::insert_account(&mut self.db.cache, address, account);

        if let Ok(frame) = self.current_call_frame_mut() {
            frame
                .cache_backup
                .entry(address)
                .or_insert_with(|| previous_account.as_ref().map(|account| (*account).clone()));
        }
```

With the changes here, we now track the individual storage slots that
are modified when executing and avoid cloning the entire modified
storage. This was done by replacing the `CacheBackup` with a
`CallFrameBackup` that keeps separate track of account infos and storage
slots.

The performance benefits are noticeable mostly in very large load tests,
with block gas limits around 1 Gigagas and beyond. At that point the
load test gets around 2x faster compared to main (80 seconds down from
160s for the load test to finish, gigagas/s goes from ~0.11 to ~0.2).

This is also noticeable within flamegraphs of these load tests, as in
`main` currently there's a huge portion of it devoted to `sstore` that
disappears.

Main:

<img width="1505" alt="Screenshot 2025-05-08 at 11 20 53"
src="https://github.com/user-attachments/assets/5d7f6dbd-d4eb-42e3-bec3-b81632ec9409"
/>

This branch:

<img width="1503" alt="Screenshot 2025-05-08 at 11 22 03"
src="https://github.com/user-attachments/assets/b6e83203-f31c-46fc-8434-998e585cabbf"
/>

While most instances of deployed ethrex will probably not feature such
large gas limits on blocks, this change should improve syncing times, as
there we execute batches of 1024 blocks at a time, which is functionally
equivalent to executing one very large block when it comes to the
behaviour of this code. Trying out the changes while syncing in Holesky,
I noticed an improvement of around the magnitude above (~2x), though of
course it is highly variable and dependent on which blocks get executed.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>