feat(l2): prove withdrawals (#2668)
**Motivation**

We want to prove the L2 withdrawals in our prover.

**Description**

- Add to `ProgramInput` and `ProverInputData` the field
`withdrawals_merkle_root` the merkle root that is created by merkelizing
all the withdrawals from a batch of blocks to send to the prover
- Inside the prover add logic to for every batch:
- Gather the withdrawals hashes for each block from the block's
transactions.
  - Get the root of the Merkle tree from these hashes
- Compare our resulting Merkle root with the incoming from the
`ProgramInput`
- Modify the l2 integration-test so that it executes 5 withdrawals to
ensure that the merkelization is working correctly
- Added a dirty fix where `cargo check` is complaining about a missing
field in the struct `ProgramInput` because the crate `l2` enables the
feature `l2` in `zkvm-interface` but neither of those crates depend on
`ethrex-prover` so the feature isn't enabled
  - This should be fixed by completing #2662

Closes #2201

---------

Co-authored-by: Estéfano Bargas <estefano.bargas@fing.edu.uy>