refactor(l1): rename incorrect usage of `node_id` to `public_key` (node_id refactor 1/3) (#2778)
**Motivation**
Our implementation of `Node` stores the node's public key as `node_id`
which is very confusing, as the `node_id` is the keccak256 hash of the
public key. This can lead to potential bugs and discrepancies with other
implementations where node_id is indeed the keccack hash of the public
key.
For this PR I left the public key as part of the Node but corrected its
name to `public_key`, leaving all use cases as is.
I also renamed some functions that mislabeled public key as node_id to
better reflect what they do. The methods `id2pubkey` and `pubkey2id`
convert between the uncompressed (H512) and compressed (PubKey) versions
of the same data so I renamed them to `compress_pubkey` and
`decompress_pubkey`.
I also added the method `node_id` to `Node` which returns the actual
node_id (aka the keccak252 hash of the public key).
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Rename various instances of `node_id` to `public_key`
* Rename methods `id2pubkey` and `pubkey2id` to `compress_pubkey` and
`decompress_pubkey`.
* Add `Node` method `node_id`
<!-- A clear and concise general description of the changes this PR
introduces -->

**Potential Follow-Up work**
Cache node_id computation so we don't need to hash the public key on
every Kademlia table operation (#2786 + #2789 )
<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2774