fix(core): timeout due to node inactivity instead of total load test time (#2530)
Changes:

- Timeout is now smarter. Instead of waiting a fixed amount of time
(e.g. 10 minutes) for the whole load test to happen, which is a bit
unpredictable, the load test waits at most 1 minute (configurable) of
no-updates from the node. This way it's less machine dependent and more
based on responsiveness.
- load-test-ci.json is fixed to be similar to perf-ci.json, but in
prague and with the system smart contracts from l1-dev.json deployed.
- logs are re-added.
- Readme si fixed.
- Re-add flamegraph reporter to CI so they are generated on every push.

Closes #2522