chore(l1): fix remaining EIP-7002 and EIP-7251 ef tests (#2738)
**Motivation**

There are tests from EIPs 7002 and 7251 that are being skipped on LEVM
and REVM.

**Description**

According to
[EIP-7002](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-7002.md)
and
[EIP-7251](https://github.com/ethereum/EIPs/blob/master/EIPS/eip-7251.md),
intrinsic gas does not count on system calls defined in this especific
EIPs. In order to avoid introducing complexity elsewhere in the codebase
(such as in intrinsic gas computation), system call contexts were
updated to include an extra 21,000 gas, the base cost of any
transaction.