fix(l1): add deposit request layout validations + return invalid deposit request error (#2832)
**Motivation**
Currently, when we fail to parse a deposit request we simply ignore it
and keep the rest of the deposits, relying on the request hash check
afterwards to notice the missing deposit request. This PR handles the
error earlier and returns the appropriate `InvalidDepositRequest Error`.
This will provide better debugging information and also more accurate
testing via tools such as `execution-spec-tests` which rely on specific
error returns.
We also were not correctly validating the layout according to the
[EIP](https://eips.ethereum.org/EIPS/eip-6110), as we were only checking
the total size and not the size and offset of each request field
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Check that the full layout of deposit requests is valid (aka the
internal sizes and offsets of the encoded data)
* Handle errors when parsing deposit requests
* Check log topic matches deposit topic before parsing a request as a
deposit request
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Allows us to address review comment made on execution-specs-test PR
https://github.com/ethereum/execution-spec-tests/pull/1607 + also closes
#2132