chore(l1,l2): remove double Arc and Mutex from metrics (#2847)
**Motivation**

The underlying Gauges are already thread safe and behind Arcs
internally, so the used Arc and Mutex wrapper were useless overhead.

<!-- Why does this pull request exist? What are its goals? -->

The types in the library derive from

```
pub struct GenericCounter<P: Atomic> {
    v: Arc<Value<P>>,
}
```

Which is clone safe, furthermore P is atomic so it doesnt need a lock.

**Description**

Remove unused Mutex and Arc

Closes #issue_number