docs(levm): mention solidity compiler dependency for levm benchmark tool (#2906)
**Motivation**
Following the `rev_comparison` levm bench tool's README to run the tool
results in an error as the solidity compiler is not installed. This is
not mentioned as a dependency, so this PR updates the README to mention
that solidity compiler is required to use the tool and how to install
it.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Mention solidity compiler dependency + how to install it on levm's
`rev_comparison` benchmark tool README
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number