feat(l1): add eth to enr (#2654)
**Motivation**

*Depends on #2679*

The `eth` pair was not implemented in the current node record struct. 

Using this information allow us to discard incompatible nodes faster
when trying to connect to a new peer

**Description**

The fork_id struct is updated every time a ENRresquest/ENRresponse msg
is being received.

The `eth` pair contains a single element list, which is a ForkId
element. It's encoded/decoded using the default RLP procedure.

The P2P TCP connections starts only when the ENR is validated. The logic
of this was changed, before when a pong was received (after our ping) a
new TCP connection was established. Now, the ENR pairs have to be
validated in order to start a new TCP connection.

When exchanging the ENRrequest/ENRresponse messages the `eth` pair is
now included

It can be tested starting a new node with debug level:
```
cargo run --bin ethrex -- --network test_data/genesis-kurtosis.json --log.level=debug
```

And connecting a new peer using the following command:
```
cargo run --bin ethrex -- \
--network ./test_data/genesis-kurtosis.json \
--bootnodes=$(curl -s http://localhost:8545 \
-X POST \
-H "Content-Type: application/json" \
--data '{"jsonrpc":"2.0","method":"admin_nodeInfo","params":[],"id":1}' \
| jq -r '.result.enode') \
--datadir=ethrex_c \
--authrpc.port=8553 \
--http.port=8547 \
--p2p.port=30388 \
--discovery.port=30310
```

A debug msg can be seen once the `eth` pair is validated

Also a new test has been done to test this new validation adding a node
with a correct forkId and a node with an invalid forkId.

For running the tests, it was needed to initialize the db so the forkId
struct could be built from those values.

[ETH enr
entry](https://github.com/ethereum/devp2p/blob/master/enr-entries/eth.md)
[EIP-2124](https://eips.ethereum.org/EIPS/eip-2124)
[Ethrex
Docs](https://github.com/lambdaclass/ethrex/blob/main/crates/networking/docs/Network.md)

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #1799