fix(levm): return EVM error when block hash isn't found in the database (#2940)
**Motivation**

- Return error when block isn't found in the database in LEVM.

**Description**

- `get_block_hash()` now returns directly the block hash instead of an
Option. If the block hash wasn't found in the DB, return an error
(instead of returning a `None` value)

**Extra details**
Before we weren't returning an error when the block looked for in the
database wasn't found. Instead, in LEVM we were just pushing 0 to the
stack, which is completely wrong, because the block should've been in
the database in the first place.

This was discovered when trying to import blocks from Hoodi testnet.
Previous Error Message:
`WARN ethrex::cli: Failed to add block 1579 with hash
0x6b910c7ee94818d5b7a6422f981159ccf438aac0f0eed20810b2a783d7c05f4d:
Invalid Block: Gas used doesn't match value in header`.
Current Error Message:
`WARN ethrex::cli: Failed to add block 1579 with hash
0x6b910c7ee94818d5b7a6422f981159ccf438aac0f0eed20810b2a783d7c05f4d: EVM
error: Database access error: DB error: Block header not found for block
number 1550.`