fix(l1): validate received blocks from peers (#2658)
**Motivation**

While syncing on Holesky, we noticed that some peers were providing us
with empty block bodies. This made the syncing loop fail in different
occasions, which ultimately lead to a stop in the syncing process.
Instead of failing, we want to validate the received headers and bodies
and discard the peer and retry if they are not valid.

**Description**

This PR makes the following changes:

* Add a new validation `validate_block_body` to check if a body is
valid, agains the corresponding header
* Add a simple header validation to the peer handler request headers, to
make sure that the received headers conform a chain from the current
head
* Call the `validate_block_body` function from the peer handle validat
and request block bodies, to check that the received block bodies are
valid.
* Modify the `PeerChannel` API to return the `peer_id` on the
`request_block_headers` and `request_block_bodies` functions.
* Modify the `PeerChannel` API to add a function `remove_peer` to remove
a peer by its `peer_id`.
* Remove the peer that provided us with the headers or bodies on the
syncing loop if they are invalid

Closes #2766

---------

Co-authored-by: Julian Ventura <julian.ventura@lambdaclass.com>
Co-authored-by: Rodrigo Oliveri <rodrigooliveri10@gmail.com>
Co-authored-by: SDartayet <sofiadartayet@gmail.com>
Co-authored-by: SDartayet <44068466+SDartayet@users.noreply.github.com>