ci(l1): added hive client list to use ethrex as baseimage (#2955)
**Motivation**

Locally built ethrexs aren't getting picked up by hive, because it
searchs for the uploaded version with the ghcr.io tag. This change
should make it search for the version tagged only with "ethrex". This
affects PRs CI too, as they would download the ethrex version that is
effectively the one from main.

**Description**

* Added a client-file to Hive command, making it build with the local
one.
* Changes the Makefile to use that client-file.
* Changes to the CI to use that branch,

---------

Co-authored-by: Mechardo <30327624+mechanix97@users.noreply.github.com>