chore(l1): remove libsecp256k1 and use secp256k1 (sp1) instead (#2956)
**Motivation**

Ethrex is composed of several smaller projects and each of them uses a
different library for secp256k1 cryptography.

This is not a problem right now because the usage of some of these
libraries is gated by feature flags (i.e. `l2` flag). But if we aim to
reduce the amount of feature flags used across the project we will need
to use a single lib to avoid future errors and name clashes.

**Description**

This PR removes the `libsecp256k1` crate from the workspace, in favour
of the one required by `SP1` (the patched `secp25k1` crate).

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Manuel Iñaki Bilbao <manuel.bilbao@lambdaclass.com>