feat(levm): set block custom block batch size and sync start with env variables (#2812)
**Motivation**

Being able to make the holesky sync testing more automated and
streamlined

**Description**

This PR introduces a feature flag that's off by default, but can be
turned on to customize some options for performance testing. It also
introduces some checks for environmental variables that could be set to
customize some aspects of the syncing when testing client performance.

Closes [#2811](https://github.com/lambdaclass/ethrex/issues/2811)

---------

Co-authored-by: Rodrigo Oliveri <rodrigooliveri10@gmail.com>