fix(l1): keep track of previously estimated gas tips (#2858)
**Motivation**
This aims to solve the case where a null gas tip is returned by
`eth_maxPriorityFeePerGas` endpoint when the blocks used for tip
estimation are empty (aka don't have transactions to obtain the tip
from).
Looking at Geth's implementation, I saw that for these cases [the
previously calculated gas tip is used in place of the empty block's
transactions'
tip](https://github.com/ethereum/go-ethereum/blob/368e16f39d6c7e5cce72a92ec289adbfbaed4854/eth/gasprice/gasprice.go#L195).
This PR aims to implement this logic by keeping track of the last
estimated gas tip and using it in the case where gas tips are collected
from the queried blocks. It also keeps track of the latest hash for
which we estimated the tip so we don't need to estimate more than once
for a given block.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add `GasEstimator` in place of the standalone `estimate_gas_tip` so we
can save and use data from previous estimations when estimating gas tips
* `estimate_gas_tip` no longer returns an `Option`
* Avoid estimating gas tips more than once for a single latest block
* Use previously estimated gas tip when we cannot obtain gas tip values
from the latest blocks (aka all queried blocks are empty of
transactions)
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->
**Comments**
The diff of `estimate_gas_tip` is not visible due to it being moved
inside the impl block and its file being renamed, but the main logic was
left unchanged, the only changes are at the top (using precomputed value
if the tip was already estimated for that block) and at the bottom
(using the previously computed value if no tips can be obtained + saving
current values)

Closes #2682