fix(levm): edge case error when creating contract fails (#2984)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- When syncing with holesky in batches there's an error particularly
when a Create internal transaction reverts, this is because we are
mishandling the cache. This aims to fix it.

**Description**

- Created transfer function to make transfers more clear
- Make changes that should be reverted if tx reverts after pushing new
callframe to callstack so that these are reverted when that callframe
reverts. This allows deleting manual reversion, which is error prone.
- Restore cache state in 2 places:
  - In handle_return when internal transaction reverts
  - When the external transaction (the first callframe) reverts
- The substate restore is done only when an opcode fails

Note: Code quality can be improved to make things be nicer and it will
be handled in the future, but for now this fixes the bug in the most
clear way possible.


Closes #2992