feat(core): ethrex_replay transaction replay (#2982)
**Motivation**

Sometimes we want to debug a specific transaction, and see the state
changes it performed.

**Description**

This moves the (l1-only) block setup to it's own function so it can be
called from run_tx.

It executes every transaction leading up to it in the block, ensuring to
only keep minimal state.

Makes progress towards #2913