fix(l2): store batch before sending (#2978)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
There's a bug where the L1 committer sends a transaction and it's not
confirmed within the expected period, so the committer assumes a failure
and returns an error. Then, if the transaction is eventually executed,
the batch is never stored. This provokes that the proof coordinator
doesn't have the batch to send to the prover, so the batch is never
verified.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Store the batch before sending the commitment transaction. Then, in the
proof sender, check that the next batch to verify corresponds with the
last committed block.

<!-- Link to issues: Resolves #111, Resolves #222 -->

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>