fix(l2): fix transactions getting stuck in the mempool forever due to low nonce (#3035)
**Motivation**

The changes make the payload builder check if a transaction has a stale
nonce and discard it if so. This is not a full solution as I'm pretty
certain there could be other reasons for transactions to get stuck in
the mempool, though this is the only one we have encountered so far. In
the future we may want to have some more involved logic to eventually
evict txs from the mempool but for now this suffices

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number

---------

Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>