refactor(l2): review l2 clis args (#2744)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

Some of the existing flags in the different L2 CLIs inherited default
values from their previous implementation (toml files). These values
were meant to facilitate dev deployment.

As we're preparing ethrex for a more production-like environment,
keeping some options with default becomes kind of an issue because the
CLI doesn't ask the user for these, leading to incorrect deployments and
so on.

The correct approach for this would be for the CLI to require some
flags, and to update the dev tooling commands (like our makefile) to
pass the default values. This way, day-to-day devs devex is not impacted
(they still deploy fast and with default values they don't care about),
and prod teams like the devops team can be sure about which
configuration is required for a prod-like deployment (not even for
production, but also for testnet, you wouldn't want to burn testnet
tokens in vain).

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Manuel Iñaki Bilbao <manuel.bilbao@lambdaclass.com>
Co-authored-by: Francisco Xavier Gauna <francisco.gauna@lambdaclass.com>