chore(l1): add total gas used to execution metrics when syncing (#3067)
**Motivation**

Looking at execution metrics while syncing I found myself wanting to see
not just total transactions of the batch, but also total gas. The total
number of transactions can be misleading since they could be a small
number of very heavy txs, or they could just be a small number of
transfers, and that changes a great deal about what is being executed.
Total gas is a better proxy to how heavy a batch is.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number