fix(l1): client discarding repeated tx when received from different peers (#3092)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
We noticed that when the same tx was received from two different peers,
the first one will be OK but the second one will be discarded and no
been broadcasted.

**Description**

The fix is to check if the tx is the same when validating the nonce in
the mempool.

Also improves the error msg
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number