refactor(levm): improve errors (#3055)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
- The error structure that we had before was messy and it needed to be
refactored. Created some extra issue for tackling some particular
changes that weren't handled in this PR.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
I propose a new structure for our errors, before we had a lot of stuff
inside VMError and it was very cluttered.
The new structure of VMError is:
```rust
#[derive(Debug, Clone, PartialEq, Eq, thiserror::Error, Serialize, Deserialize, Display)]
pub enum VMError {
    /// Errors that break execution, they shouldn't ever happen. Contains subcategory `DatabaseError`.
    Internal(#[from] InternalError),
    /// Returned when a transaction doesn't pass all validations before executing.
    TxValidation(#[from] TxValidationError),
    /// Errors contemplated by the EVM, they revert and consume all gas of the current context.
    ExceptionalHalt(#[from] ExceptionalHalt),
    /// Revert Opcode called. It behaves like ExceptionalHalt, except it doesn't consume all gas left.
    RevertOpcode,
}
```
This PR also:
- Deletes a lot of errors that we weren't using
- Renames some errors to shorter or more accurate names, changing it's
descriptions if necessary
- Removes `OutOfGasError` enum (that had multiple confusing variants)
and replaces it for a simple `OutOfGas` value.
- Transform some errors into Internal because they should've been that
kind in the first place. Like `NonceOverflow` for example (this
shouldn't revert execution because it shouldn't even happen, we do
validations before incrementing it) or `MemorySizeOverflow`.
- We had multiple types of errors for the same thing sometimes, this
happened for example with `AccountNotFound` and
`AccountShouldHaveBeenCached`, I opted to remove the latter.
- Now functions/methods are more precise about what kind of error they
return. If a function only returns Internal Errors I changed it's error
type to `InternalError` so that it's behavior is clear (This is useful
when the programmer wants to see if a method could revert execution or
not).
- Removes (alongside other errors) `UndefinedState` from
`InternalError`. This was an error type that we used when we didn't know
what kind of error to return.
- Makes usage of `thiserror` crate, that lets us return just the error
that we want to return without needing to wrap it into the error that
the function returns. It handles conversions for us when using `?`,
otherwise we use `.into()`.

Opened issues #3062 and #3063 for improving even more the error
handling.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #2886

---------

Co-authored-by: fmoletta <fedemoletta@hotmail.com>
Co-authored-by: fmoletta <99273364+fmoletta@users.noreply.github.com>
Co-authored-by: Tomás Paradelo <112426153+tomip01@users.noreply.github.com>