refactor(l2): moved websigning logic to own crate (#3081)
**Motivation**

This PR aims to do a simple fix. In the web3signer branch we have an
imports issue, where loading the reqwest package conflicts with sp1.
This is the simplest fix, where we move only the websigning logic to
it's own crate guarded by a feature flag. As such, reqwest is not
imported in L1 with sp1. This should be changed in the future, once the
signing gets properly reformatted to be part of the SDK.

**Description**

- Created a websigner crate with a single function encapsulating the
websign logic.
- Added a feature flag to common/types, to not import the websign logic
when needed.
- Made the RemoteSigner class function only with the feature flag to
sign.

Progress for #2714 
Closes #2922