fix(l1): peer scoring to avoid syncing hanging in holesky (#3042)
**Motivation**

Syncing where being hanged in a loop, not being able to proceed at the
bodies request stage

**Description**

The issue was related to an early return of an error when receiving no
bodies from a peer (after the retries). This happen even if some peer
was able to answer but we exhausted the retries with those who wont. We
tried a couple of solutions:

- **Remove the error and just break with the accumulated bodies:** This
proved to hurt the performance and just push the issue to the next
iteration which also failed to accumulate the whole amount of bodies
- **Remove the peers that doesn't answer the bodies:** This proved to
still fail when we end up with just one good active peer and for some
reason it doesn't answer. In that case we removed it also and were
unable to get more peers.
- **Remove the peer that doesn't answer and increase the timeouts:** It
was possible that the issue with the good peer was related to the short
timeout of 5 seconds, increasing this proved to be better thatn te
previous solution but after a couple of houser the issue of having no
peers still remained.

**Current solution:**

Now we implemented a initial version of a simple peer-scoring. This just
adds a new score field to the peer data and use it to weight the peers
when choosing them (instead of the previous random selection). In both
headers and bodies request we penalize and reward peers given their
responses. This could be further improved in other PRs (#3042) but it's
enough for the current issue. After running the whole night 3 different
times syncing from holesky's 3.6m there were no issues in sight! We also
maintained the increased peer timeout from 5 to 15 secs.

Additionally I removed the marking of the sync head as invalid,
something that shouldn't be needed in syncing (we are already marking
the failing one, which could also be an issue but is tracked in #2767)

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3020