refactor(levm): remove contemplations for empty accounts in trie (#3131)
**Motivation**

- In post-Merge Mainnet there aren't any empty accounts in the trie.
They were all cleared, so we don't need to implement the logic for
clearing those accounts that were just in the trie because of a "bug"
previous to Homestead fork. For more information see
[EIP-7523](https://eips.ethereum.org/EIPS/eip-7523).

**Description**

- Removes `account_exists()` method from our Database trait in LEVM, as
we don't need that information anymore.
- We also remove some unused methods that `account_exists()` from RpcDB
was using only there.
- Renames `touched_accounts` to `accessed_addresses`, which is the name
it should've had in the first place. It's just that we didn't fully
understand the difference between these two concepts and treated them as
being the same. Now that we do understand it we don't care about keeping
track of touched accounts but we do care about accessed addresses.