fix(l2): remove payment to Coinbase in `PrivilegedTransaction` (#3132)
**Motivation**

We were paying a `priorityFee` to `Coinbase` in
`PrivilegedL2Transactions`, resulting in an L2 total balance higher than
the amount locked in the `CommonBridge` contract on L1.

**Description**

- Removes the `Coinbase` payment in `PrivilegedL2Transactions`.
- Adds a check to the `l2_integration_test` to assert that the L2 total
balance is `<=` than the amount in the bridge contract.
- In order to calculate the L2 total balance, we need to track all
addresses that have received a deposit and retrieve their balances at
the end of all tests. To achieve this, this PR places all tests under
`l2_integration_test`, ensuring the test order and collecting the
necessary addresses.

Closes #3079