ci(l1): add ethrex_only localnet (#3048)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
Now that we are able to run a localnet with only ethrex instances add
one and test it with assertoor playbooks

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
The command `make localnet-assertoor-ethrex-only` was added for running
a localnet with only ethrex instances
Also a `network_params_ethrex_only.yaml` file was created to configure
the assertoor tests and was included into the l1 github workflow. The
makefile command uses this file as its configuration.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3008

---------

Co-authored-by: SDartayet <44068466+SDartayet@users.noreply.github.com>