docs(l2): move L2 docs and quickstart guide to mdbook (#3171)
**Motivation**

Our documentation is scattered around the repo. We want it to be easy to
find.

**Description**

This PR puts L2 documentation under `docs/l2` and the developer
quickstart guide under `docs/developers`.

It also adds `linkcheck` as an optional renderer, which checks for
broken links and fixes a stale part of the L2 prover documentation.

Note that I also had to cherry-pick
67ce6262178c7a37dc8328536bd9e5f34abcc9f1 from #3163

This is the current structure:

<img width="293" alt="Screenshot 2025-06-13 at 18 46 03"
src="https://github.com/user-attachments/assets/efb29ab0-b79d-4106-8e27-5c5d60fb6287"
/>