chore(l1): match blockchain tests obtained error with expected error (#3113)
**Motivation**

Include the exception comparison in the blockchain test runner, as it
could catch some inconsistencies.

**Description**

This pr introduces an Include BlockChainExecutionError and a
deserializer function to parse the expected errors from the ef tests, as
well as a exception_is_expected function, just like in
[levm](https://github.com/lambdaclass/ethrex/blob/main/cmd/ef_tests/state/runner/levm_runner.rs#L258)'s
test runner, to check if the error obtained matches the error expected.

Closes #290

---------

Co-authored-by: SDartayet <44068466+SDartayet@users.noreply.github.com>