chore(l2): change default dev-mode to false (#3214)
**Motivation**

The default value of `proof-coordinator.dev-mode` is set to true. This
means the only way to set it to false is through the environment
variable `ETHREX_PROOF_COORDINATOR_DEV_MODE`. This is also inconsistent
with the rest of the parameters, where we set dev values only in the
Makefile.

**Description**

Changes the default value of `dev-mode` to false.

Closes None