ci(l1): have failed tests output on the console (#3150)
**Motivation**

The LEVM CI workflow in pr-main_levm.yaml that runs EF state tests
should fail with an exit code if a test fails.

**Description**
This pr introduces a new `EFTestRunnerError::TestsFailed` error to use
when there's a report of a test failing. This error is thrown under the
`summary` flag, which is the one used in the target the CI job executes:
`make run-evm-ef-tests-ci`. So whenever there is any failing tests, the
introduced code should print the EFTestReport and then finish with the
`EFTestRunnerError::TestsFailed` error.

Note: The `summary` flag is used in other targets as well, so the
previously described behavior is being implemented for other targets
too.

The ef-test-main job in pr-main_levm has also been refactored, I dropped
steps "Check EF-TESTS from Paris to Prague is 100%" and "Check EF-TESTS
status is 100%" since now in the case any test fails, the CI job exits
with an error code and outputs the failing tests in the console.

In this pr there are some commits with a hardcoded error with the
intentions of having the LEVM CI workflow fail on purpose and check the
console output is the one expected.
[Here](https://github.com/lambdaclass/ethrex/actions/runs/15738130731/job/44356244936)
is a failing workflow execution under this circumstances to see. (The
underscore line above "Failed tests" was removed on a later commit.)

Closes #2887