chore(l1): add metrics port to ethrex client (#3237)
**Motivation**

When running a localnet with kurtosis the ethrex client wasn't exposing
a metrics port.

**Description**

To expose the metrics port, the ETHEREUM_PACKAGE_REVISION in the
Makefile was updated to the latest commit in our fork of
ethereum-package. Additionally, the metrics feature flag was enabled
when building the Docker image (without it, metrics won't work).
The ethereum_metrics_exporter_enabled setting was also enabled for all
participants in the ethrex-only localnet.
With these changes, we are now able to use metrics with ethrex clients.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3213