chore(l2): remove nested workspace (#3277)
**Motivation**

When installing with `cargo install --git
https://github.com/lambdaclass/ethrex.git ethrex`, I was getting weird
errors from the `zkvm_interface` crate. Looking into that, I found that
we have a nested workspace in our repo. This is not supported, and hence
should be avoided.

**Description**

This PR removes the second workspace. I also tried keeping some of the
optimization options for the package in
4635cfec948a12e52d6a61ec317781845714838c, but the `lto` option was
giving me problems. In case we want to re-add them, the other options
can be added directly to the main `Cargo.toml` instead of a
`config.toml`.