feat(l1): running localnet with client comparisions (#3221)
**Motivation**

We want to be able to spin up a local network with 3 nodes,
ethrex[levm], ethrex[revm] and reth an be able to compare throughput
between them on different spamoor configurations.

**Description**

This PR add a new make target that spin-up a localnet with ethrex[levm],
ethrex[revm] and reth, enable the metrics exporter, and prepare the
dashboard to be able to be used in arbitrary datasources, both imported
or provisioned, automatically setting a datasource variable to the
default prometheus.