test(l1): check if vm state reverts correctly on error (#3198)
**Motivation**

Per the instructions of the [ethereum state test
docs](https://eest.ethereum.org/v3.0.0/consuming_tests/state_test/), we
should be reverting to the pre-state when the execution throws an
exception. Levm does this, but it's not asserted in the test runner in
the case an exception is expected.

**Description**

This pr introduces a new error to check if the state was reverted
correctly in the case an exception must occur, or throw error otherwise.
To check if the state was correctly reverted I'm using the post state
hash from the tests and comparing it with the hash of the account's
latest state recorded in the db.

Closes #2604