refactor(l1): `SIGTERM` handling (#3288)
**Motivation**
Stopping an L1 Docker container sends a `SIGTERM` to the node process,
but since the node doesn't handle this signal, it takes 10 seconds for
the `SIGKILL` to force the shutdown.
Some temporary fixes were introduced to address this issue, but the
correct solution is to handle the `SIGTERM` signal.


**Description**
This PR adds a `SIGTERM` handler to allow graceful shutdown without the
10-second delay.

It also reverts the temporary fixes that were introduced in previous PRs
(https://github.com/lambdaclass/ethrex/commit/8f6958500db6b0610e3693f117679fccaee672b9
and
https://github.com/lambdaclass/ethrex/commit/68d2ce4992508aa80c0f51206d5b93ee02d10cf8)
in favor of this solution.

Note that when the node is syncing, shutdown isn’t immediate, as the
node waits for the current batch to complete.
To handle this, the `CancellationToken` is propagated and checked before
processing each block, which allows the node to terminate immediately,
without waiting for the whole batch to complete.

Closes #2944
Closes #3236