docs(l1): fix readme links and improve L1 documentation (#3340)
**Motivation**

The current readme links to the unrendered documentation instead of our
hosted book. Also, the general landing page doesn't have any links or
pointers on where to go next, while the L1 landing page is empty.

**Description**

This PR addresses the previous issues, adding some content to the L1
landing page and generally cleaning up the docs. It also merges the two
documentation sections in the readme and updates links to point to
docs.ethrex.xyz