chore(core): add install script (#3273)
**Description**

This PR adds an installation script with readme instructions on how to
quickly set up a local L1 with `ethrex`, without having to clone the
repo.

The idea is to extend this script once the L2 can be more easily
deployed. Right now, it requires installing two more binaries and
downloading multiple contracts (either the source or compiled
artifacts), which makes this more complex than it needs to be.

---------

Co-authored-by: Javier Chatruc <jrchatruc@gmail.com>