chore(levm): improve levm vs revm bench (#3255)
**Motivation**

Currently the benchmark comparisson shown for levm is not very useful
since there is a lot of information distributed along different tables
and it's not easy to see the change.

**Description**

This pr introduces the following changes:

- Instead of having the tables compare levm and revm, now they compare
between pr's obtained mean time and main' obtained mean time for both
vms. To do this we modify the `run_benchmark_ci` function that the
target runned by the ci calls to have the benchmark comparison run for
all cases.

- If nothing changed between pr and main, no message is printed, using
as a margin of error a porcentual difference higher than 10%.

- If something changed, only output the tests where there was a change,
nothing is printed if the test that stayed the same.

The tables always show the obtained metrics in the same order despite
the fact one of them is not shown:

| Command | 
|--------|
|`main_revm_`|
|`main_levm_`|
|`pr_revm_`|
|`pr_levm_`|

For example, in the case you do an optimization in levm that improves
factorial but does nothing for revm, the table would look like this:
| Command | Mean [ms] | Min [ms] | Max [ms] | Relative |
|--------|--------|--------|--------|--------|
| `main_revm_Factorial` | 138.1 ± 0.5 | 137.4 | 139.2 | 1.00 |
| `main_levm_Factorial` | 326.4 ± 6.3 | 321.4 | 340.2 | 2.36 ± 0.05 |
| `pr_levm_Factorial` | 223.8 ± 6.3 | 216.4 | 234.1 | 2.04 ± 0.05 |

Closes #3254

---------

Co-authored-by: cdiielsi <49721261+cdiielsi@users.noreply.github.com>
Co-authored-by: Camila Di Ielsi <camila.diielsi@lambdaclass.com>