feat(l2): value sending in privileged transactions (#3320)
**Motivation**

We want to be able to send ("deposit") value-carrying privileged
transactions, for example to do forced withdrawals.

**Description**

This PR allows users to specify the transaction value while sending
privileged transactions.

For regular deposits, the bridge calls it's L2 version and relies on the
L2 hook allowing the bridge address to send without having funds. It
calls it's own L2 version, which handles failed transfers by making a
withdrawal.

For transactions, since they can't fail they are instead made to revert
when there aren't enough funds to cover the operation.

Closes #3290, closes #3291

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>