feat(l2): exchange commit hash in node-prover communication (#3339)
**Motivation**

We want to prevent a divergence between the code that is running in the
L2 node and the prover.

**Description**

- Updates the client version to use `GIT_BRANCH` and `GIT_SHA` instead
of `RUSTC_COMMIT_HASH`.
- Adds a `build.rs` script for both the node and prover, using
`vergen_git2` to export the git env vars.
- Adds a `code_version` field to the `BatchRequest` message.
- Introduces a new `ProofData` message: `InvalidCodeVersion`.

You can create an empty commit with:

```bash
git commit --allow-empty -m "empty commit"
```

Then run the node and the prover using different commits.

> [!WARNING]
> Remember to run `make build-prover` whenever you change the commit

Closes #3311