refactor(l2): refactor privileged transactions (#3365)
**Motivation**

We want to unify the terminology used to refer to sending privileged
transactions to the L2, since they are not just deposits.

Also, the `DepositInitiated` (now `PrivilegedTxSent`) event must be
cleaned up: `l2MintTxHash` is irrelevant since it can be recomputed, and
it doesn't make sense to index `amount` but not `from`.

Some clean up (for example, removing `recipient`) was already done in
#3320.

**Description**

* Removes the l2 transaction from the deposit event
* Adds `indexed` to `from` and removes it from `address`
* Renames deposit to 'privileged transactions' where more appropriate

Closes #3233

---------

Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>
Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>
Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>