refactor(levm): separate marking account as accessed from getting its info (#3440)
**Motivation**

Behavior related to marking an account as accessed (turning it from a
_cold_ address to a _warm_ address) was attached to obtaining its
information (method invoked `get_account`).
The goal was to separate these two actions so that accessing an account
does not always imply getting its _account_info_.

Also, this method belonged to the `GeneralizedDatabase` struct which
actually has no relation to accessed addresses record (`Substate` is
responsible for this).

**Description**

The method `access_account()` in `GeneralizedDatabase` was removed. Now,
whenever an account needs to be accessed the following is invoked:
`substate.accessed_addresses.insert(address)` that returns a boolean
that if true means address was previously cold and inserted, if false
address was previously warm and not reinserted.

Closes #2899