ci(core): add label to PR automatically and assign author (#3463)
**Motivation**

Have a more organized workflow.

**Description**

Workflow 1: assignee is set when opening a PR.

Workflow 2:
Add label automatically to the PR by looking at the title, depending on
the scope.
`l1` or `levm`: Assign `L1` or `levm` label. Which is then assigned to
`ethrex_l1` project.
`l2`: Assign `L2` label. Which is then assigned to `ethrex_l2` project.
`core`: Assign both `L1` and `L2` labels. Which then is assigned to both
projects.

If a PR already has a label then the workflow doesn't assign a new label
to it. This is helpful for when the user manually set the label and
doesn't want the automation to override what he/she did

This replaces #3449 because it's a more convenient approach and doesn't
require a PAT.