fix(l1): `RpcBlock` `uncles` field should have the hashes and not the block headers (#3245)
**Motivation**
Fix inconsistencies between our RPC outputs and the spec.
According to the spec endpoints such as `eth_getBlockByNumber` return a
block where the `uncles` field contains the hashes of the uncle blocks,
while we return the full headers.
This has not been a problem for us as we have been mainly using
post-merge blocks without uncles, but it will become a problem if we
need to export/import older blocks via rpc
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Change `uncles` field of `RpcBlock` from `Vec<BlockHeader>` to
`Vec<H256`
* (Bonus) Allow deserializing blocks without `base_fee_per_gas`
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes None