fix(l2): rename field  `from` of `PrivilegeL2Transaction` for serialization (#3487)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

The field `from` of the struct `PrivilegeL2Transaction` overlaps with
the field `from` of
[`RpcTransaction`](https://github.com/lambdaclass/ethrex/blob/main/crates/networking/rpc/types/transaction.rs#L15)
during serialization (because the field `tx` is flattened by `serde`),
making `RpcTransaction` deserializations invalid for
`PrivilegeL2Transaction`s.

**Description**

Renames the field `from` of `PrivilegeL2Transaction` (only for
serialization and deserializations) to fix the deserializations of
`RpcTransaction`s of `PrivilegeL2Transaction`.