refactor(levm): change returned error types to `InternalError` (#3322)
**Motivation**
From [#3063](https://github.com/lambdaclass/ethrex/issues/3063)

> There are various cases in which we return an error with the
ExceptionalHalt type but they actually are InternalErrors, things that
shouldn't ever happen and if they happen they should break.
This is not a critical issue since if the VM is working fine then it
won't ever enter to those cases, but it would be more precise if we
catalogued those errors as internals instead of saying that they revert
execution when they don't.

<!-- Why does this pull request exist? What are its goals? -->

**Description**

Introduces the following changes:
* Replaces `PrecompileError` with `InternalError` in those cases in
which an error is returned even though is not possible for the
instruction to fail, typically when slicing bytes whose size have been
already checked.
* Removes the error types `EvaluationError` and `DefaultError` (which
were quite generic) from `PrecompileError` and adds specific and more
descriptive error types instead (`InvalidPoint`, `PointNotInTheCurve`,
etc).
* Removes the `PrecompileError::GasConsumedOverflow` error type.



Closes #3063

---------

Co-authored-by: Jeremías Salomón <48994069+JereSalo@users.noreply.github.com>
Co-authored-by: Martin Paulucci <martin.c.paulucci@gmail.com>