ci(l1,l2): remove `core` scope and improve PR labeling workflow (#3561)
**Motivation**

- Declutter `ethrex_l1` project and remove ambiguous `core` scope in
title.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- Remove core scope because it is ambiguous. Replace it with `l1,l2`
- Merge `pr_author.yaml` and `pr_label.yaml` into one file
`pr_github_metadata.yaml`
- Change rules of labeling because of preferences in our projects:
- `ethrex_performance`: Will have PRs that have `perf` at the beginning
of the title.
  - `ethrex_l2`: Will have any PR that has in the title scope `l2`
- `ethrex_l1`: Will have PRs that haven't been assigned to
`ethrex_performance` or `ethrex_l2` that have `l1` or `levm` in their
title.

The decisions were made according to the preferences of each team.
`ethrex_l2` project will have anything that has to do with the L2
`ethrex_l1` project will have things that touch only l1 stuff and
nothing else so that we assure they truly belong to this project. Some
PRs will be filtered out and will have to be added manually, but we
prefer that rather over the clutter of having more PRs than necessary.

Closes #3565