fix(l2): seal a batch in a single DB transaction  (#3554)
**Motivation**

When deploying ethrex L2 some errors came up that are related to the
seal_batch process not being done in a single DB transaction.

**Description**

- Move seal_batch to the `StoreEngineRollup` trait
- For sql rollup store engine
- Wrap all the DB write functions from the trait with a <name>_in_tx
that gets as an input an Option<Transaction> in case the transaction is
Some then it uses the existing transaction, and does not commit. If its
None it creates a new transaction and commits at the end of the
function.
- Modify the `SQLStore` struct to hold two instances of `Connection` one
for reads and one for writes, the write connection is protected by a
Mutex to enforce a maximum of 1 to prevent this error:
      ```
failed because of a rollup store error: Limbo Query error: SQLite
failure: `cannot start a transaction within a transaction`
      ``` 
- Use `PRAGMA journal_mode=WAL` for [better
concurrency](https://sqlite.org/wal.html#concurrency)
- For `libmdbx` , `redb` and `in-memory`
   - Implement the `seal_batch` function 
- Refactor: remove all the functions that were exposed by `store.rs` and
were only part of seal_batch to prevent its usage outside of batch
sealing.


Closes #3546