ci(l1, l2): create reusable step to install rust (#3591)
Continuation of https://github.com/lambdaclass/ethrex/pull/3318

**Motivation**

GitHub Variables are excluded from workflow runs triggered by PRs from
forks, so we need to remove this variable dependency in order for
external collaborators to send PRs and run the CI properly

**Description**

* The `Extract Rust version from rust-toolchain.toml` step (`id:
rustver`) uses `grep` and `sed` to extract the rust version from the
`rust-toolchain.toml` file that is in the root of the repository.
* The `Install Rust` step utilizes the output of the previous step to
send the version to the `toolchain` parameter
* Note that in some cases, I had to move the `Checkout` step further up
(it's also good practice to put it as high up as possible) so the
`rust-toolchain.toml` file is available to be read.

---------

Co-authored-by: Klaus Lungwitz <klaus.lungwitz@lambdaclass.com>
Co-authored-by: Javier Rodríguez Chatruc <49622509+jrchatruc@users.noreply.github.com>