chore(levm): upgrade levm vs revm bench  (#3557)
**Motivation**

- Ci generated comment is not updating.
- Using a porcentual difference higher than 10% for showing results is
to high.

**Description**

This pr tackles two minor issues. 

- It modifies the bash script that puts the comment together, so the
comment can be found and updated by the ci.
- It also changes the margin of error for which the benchmarks should be
shown. Now the benchmarks are shown for a porcentual difference higher
than 5%, otherwise you can check all the benchmarks on the Detailed
Results drop-down tab. This change is introduced with the intention of
solving issue #3462

---------

Co-authored-by: Edgar Luque <git@edgl.dev>