ci(l1,l2): rename release binaries (#3603)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->
In Ansible, server architectures are read as `x86_64` and `aarch64`, so
it's easier to download the binaries if they have that names,
considering we didn't choose current names on any purpose.

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
Rename `x86-64` with `x86_64` and `amd64` with `aarch64`. Also made some
refactor to simplify the strategies matrix and move big prover/replay
`if`s to workflow level instead of shell

<!-- Link to issues: Resolves #111, Resolves #222 -->