test(l1): add tests with different clients (#3324)
**Motivation**

We were testing ethrex with lighthouse only, we should also be testing
it with other cl
<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->

Created a localnet with 3 nodes all using ethrex as its el and a
different cl.

The consensus client being used are:

- Lighthouse
- Prysm
- Teku

Assertoor is used to run tests, it checks the chain stability and that
all nodes are able to propose both BLOBS and EOA transactions.

Dora is also available to explore the network.

The `ethrex-only` localnet and jobs were replaced.

Use `make localnet-assertoor-different-cl` to run the localnet.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3140