fix(levm): ignore DB storage values for destroyed accounts (#3617)
**Motivation**
When executing blocks in batches an account may be destroyed and created
again within the same batch. This can lead to errors as we might try to
load a storage value from the DB (such as in an `SLOAD`) that doesn't
exist in the newly created account but that used to be part of the now
destroyed account, leading to the incorrect value being loaded.
This was detected on sepolia testnet block range 3302786-3302799 where a
an account was destructed via `SELFDESTRUCT` and then created 6 blocks
later via `CREATE`. The same transaction that created it then performed
an `SSTORE` which was charged the default fee (100 gas) as the stored
key and value matched the ones in the previously destroyed storage
instead of charging the storage creation fee (2000 gas). The value was
previously fetched from the DB by an `SLOAD` operation.
This PR solves this issue by first checking if the account was destroyed
before looking up a storage value in the DB (The `Store`). If an account
was destroyed then whatever was stored in the DB is no longer valid, so
we return the default value (as we would do if the key doesn't exist)
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* (`levm` crate)`GeneralizedDatabase::get_value_from_database`: check if
the account was destroyed before querying the DB. If the account was
destroyed return default value
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number