feat(l1): peer scoring for snap requests (#3334)
**Motivation**
Integrate and adapt the peer scoring introduced by #2115 for snap
requests.
For eth requests, we consider failure to return requested data as a peer
failure, but with snap the data we request is not guaranteed to be
available (as it might have become stale during the sync cycle) so we
cannot asume that an empty response is a bad response that should be
penalized. For snap requests this PR collects the ids of the peers we
attempted to request data from, and once we get a successful peer
response we confirm that the data was indeed available and reward the
responsive peer while penalizing the previous unresponsive peers
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Collect ids of peers on each snap request retry and penalize and
reward peers accordingly upon a successful peer response
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3118