feat(l2): embed proxy contract in the SDK (#3443)
**Description**

This PR adds a `build.rs` build script to the SDK to embed the
`ERC1967Proxy` contract as a constant. As part of this, it also moves
the functions for downloading dependencies and compiling contracts to
another crate, since we need to use them inside the build script.

Change list:

- [x] Added build script
- [x] Added installation of `solc` for compiling in the CI
- [x] Updated dockerfiles to install solc before compiling
- [x] Updated `service.nix` to clone dependencies before building.
- [x] Removed `ERC1967Proxy` compilation steps from the Deployer.

Related to #3380

---------

Co-authored-by: avilagaston9 <gaston.avila@lambdaclass.com>
Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>