fix(l1): remove "p2p" capability from supported capabilities list (#3571)
**Motivation**

Even if the spec calls it the "p2p" capability, it isn't included in the
`capabilities` array by our peers, but implicitly in the message, with
the version specified in the `protocolVersion` field.

**Description**

This PR removes the "p2p" capability from the capability list exchanged
through Hello RLPx messages. It also removes the unused code related to
it.