chore(l1): include Merkle step in metrics logs (#3630)
In preparation for performance analysis tooling, we need to include
Merkle-tree specific measures.
Rigorously speaking, this includes most of the data preparation before
committing to the DB. The code assumes Merkle-tree operations to be the
bottleneck there.

The notebook part will be handled in a separate PR based on this branch,
so we can keep the logic in it simpler by only handling one version of
the logs at a time. It also should make it easier to review.

Based on #3274
Coauthored-by: @Arkenan

Part of: #3331