fix(l2): estimate gas in call_to_contract_with_deposit (#3734)
**Motivation**

The CI is failing on
[main](https://github.com/lambdaclass/ethrex/actions/runs/16376083320/job/46276248732)
with the following error:

```
thread 'l2_integration_test' panicked at crates/l2/tests/tests.rs:1604:65:
called `Option::unwrap()` on a `None` value
```

This is because we were using a hardcoded `gas_limit` for the
`l1_to_l2_tx` in the `call_to_contract_with_deposit` test, and sometimes
the tx fails due to the gas limit being exceeded. Then, the expected
logs of are never created.

**Description**

- Replaces the hardcoded `gas limit` with `None` to allow the SDK to
estimate the value.

Closes None