perf(l1): use rayon for recover address (#3709)
**Motivation**
This logic was originally introduced in #2268 but was mistakenly removed
in a refactor
[PR](https://github.com/lambdaclass/ethrex/pull/3082/files#diff-6ca74a0741dab646bb82b83636f9513d38a1c66b9db52dae8e20a0ec2fe6c1a3L239-L241).
We're adding it back because it improves the performance of our
benchmarks.

**Description**

Add `par_iter` to `recover_address` function 

Biggest changes
- ETH transfers (+117.96%)
- Gas-Pop (+36.35%)
- Push0 (+33.08%): Push de ceros al stack mucho más eficiente
- Timestamp (+20.26%)
- CoinBase (+17.18%)
- Caller (+15.19%)
- GasLimit (-14.33%)
- BlobHash (-12.97%)

This change affects most of the benchmark tests, but it restores logic
that had previously been part of the codebase.

closes #3725