refactor(levm): rewrite of state EF tests runner first iteration (#3642)
**Motivation**

Related issue: #3496. 

The idea is to incrementally develop a new EF Test runner (for state
tests) that can eventually replace the current one. The main goal of the
new runner is to be easy to understand and as straightforward as
possible, also making it possible to easily add any new requirement.

**How to run** 
A target in the makefile was included. You can, then, from
`ethrex/cmd/ef_tests/state/` run `make run-new-runner`. If no specific
path is passed, it will parse anything in the `./vectors` folder.
Otherwise you can do, for example:
`make run-new-runner TESTS_PATH=./vectors/GeneralStateTests/Cancun` to
specify a path.

This command assumes you have the `vectors` directory downloaded, if not
run `make download-evm-ef-tests` previously.

**Considerations**

The main changes are: 
- The new `Test` and `TestCase` structures in types. 
- The runner and parser simplified flows. 

Files that should not be reviewed as they are full or partial copies of
the original files:
- `runner_v2/deserialize.rs`
- `runner_v2/utils.rs`

This iteration excludes report-related code, option flags and other
possible test case errors to be considered that will be included later.
Checks are performed only on exceptions and root hash.