feat(l2): based P2P (#2999)
**Motivation**

This PR follows #2931 . We implement some basic functionality to
communicate L2 based nodes via P2P.

**Description**

- Add new capability to the RLPx called `Based`.
- Add new Message `NewBlock`.
  - Behaves similar to the message `Transactions`.
- Every interval we look to the new blocks produced and send them to the
peer.
- Add this message to the allowed ones to be broadcasted via the P2P
network.
- When receiving this message we implemented a queue to be able to
receive them in disorder. Once a continuos interval of blocks is in the
queue we store them in order.
- Add new message `BatchSealed`
- Every interval we look in the `store_rollup` if a new batch has been
sealed and then we send it to the peer.
- Add this message to the allowed ones to be broadcasted via the P2P
network.
- This two new messages are signed by the lead sequencer who proposed
the blocks and the batches. Every node must verify this signature
correspond to the lead sequencer
- Change `BlockFetcher` to not add a block received via the L1 if it
already has been received via P2P, and vice versa.
- Add a new `SequencingStatus`: `Syncing`. It is for nodes that are not
up to date to the last committed batch.

**How to test**

Read the `Run Locally` section from `crates/l2/based/README.md` to run 3
nodes and register 2 of them as Sequencers. It is important that you
assign different values in the nodes:
- `--http.port <PORT>`
- `--committer.l1-private-key <PRIVATE_KEY>`
- `--proof-coordinator.port <PORT>`
- `--p2p.port <P2P_PORT>`
- `--discovery.port <PORT>`

> [!TIP]
> To enrich the review, I strongly suggest you read the documentation in
`crates/l2/based/docs`.

---------

Co-authored-by: Leandro Serra <leandro.serra@lambdaclass.com>
Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: Ivan Litteri <67517699+ilitteri@users.noreply.github.com>
Co-authored-by: Avila Gastón <72628438+avilagaston9@users.noreply.github.com>
Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>
Co-authored-by: fkrause98 <fkrausear@gmail.com>
Co-authored-by: Francisco Krause Arnim <56402156+fkrause98@users.noreply.github.com>