docs(l1): move install instructions to new section and embed script one-liner (#3505)
**Motivation**

Since the install script just builds from source using a `cargo install`
one-liner, it's preferable to show that instead of having to download
and run an install script.

**Description**

This PR removes the install script, embedding the one-liner inside the
docs. It also moves the installation instructions to the book, linking
to it in the readme, and expands them with instructions on how to build
from source or download the pre-built binaries.

---------

Co-authored-by: Manuel Iñaki Bilbao <manuel.bilbao@lambdaclass.com>