ci(l1,l2): lines of code report (#3764)
**Motivation**
The lines of code CI compares directly with main, so if the branch is
not updated it gives fake results
<!-- Why does this pull request exist? What are its goals? -->

**Description**
Now the CI compares to the most common commit between the branches

**How to Test**

This PR is not up to date with main, without the changes on the CI we
would see a code report with the new changes on main.
Since this PR only changes the CI there should be no code changes on the
report, but in order to show that it works, a commit was added where one
line changed, a report was generated, and then the commit was reverted.
The old report is still present since the new run of the CI detected
that there are no changes, so the comment is not updated.
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->