perf(levm): enable lto codegen (#3830)
**Motivation**

Lto seems to improve perfomance in some benches without regressions.

A question would be, do we add it as another profile that we use
explicitly for releases or keep it as default for --release. I ask this
question because the --release flag is used in lot of testing code and
it would slow compile times of those

<img width="2560" height="2235" alt="image"
src="https://github.com/user-attachments/assets/e79fbb09-237f-4e91-a843-5f91e9d4ecbb"
/>