feat(l2): implement scrolling in monitor tables (#3717)
**Motivation**

Since not many rows fit on the screen, we want to be able to scroll each
table.

**Description**

Since stateful tables support scrolling, we just have to wire that. Also
we change the color of the box depending on the selected element to show
the user which table is selected for scrolling.

Closes #3520, closes #3521, closes #3522, closes #3523, closes #3524,
closes #3525