fix(l2): integration tests (#3851)
**Motivation**

Transactions revert sometimes because of exceeding the estimated gas
limit.

The estimation is done by executing the transaction with different gas
limits, binary searching the smallest value to which the tx does not
revert. We think this algorithm doesn't work properly when there's a big
variance in the execution gas, resulting in a too low estimation.

<!-- Why does this pull request exist? What are its goals? -->

**Description**

- adds logs to tests
- doubles the gas limit estimation of test calls
- changes the "not an Integration Test" name to make them required again
- asserts transactions succeeded (receipt.status == true)


closes #3828

---------

Co-authored-by: avilagaston9 <gaston.avila@lambdaclass.com>