perf(l1): add (sender,nonce)=>tx_hash mapping (#3865)
Some mempool operations need to query transactions by their sender and
nonce. One of those cases is to know which transaction to replace when
sender and nonce match.
Until now, we did that by iterating the whole mempool. This became a
bottleneck for the mempool, given it's quadratic to the number of txs
you want to add.
This adds an auxiliary BTreeMap from the (sender, nonce) to the
transaction's hash.
This also helps for querying the current nonce of an address.

Testing with a load test on ethrex-sync-3 (AMD Ryzen 5 3600 6-Core
Processor, 2 threads-per-core, 64GB RAM):
- Reported time went down from 84" to 71";
- Contribution of `find_tx_to_replace` to `add_transaction_to_pool` went
  down from 80% to 1.6% according to samply;
- Max throughput slightly improved, from 0.292Ggas/s to 0.362Ggas/s.

Fixes #3867

---------

Co-authored-by: Edgar <git@edgl.dev>