fix(l1):  use p2p's `sync-test` feature in crate's `sync-test` feature (#3889)
**Motivation**
`ethrex-p2p` crate has code gated behind a `sync-test` feature, but this
feature is not enabled within the crate's `sync-test` feature, causing
this code not to be enabled when running ethrex with `sync-test`
feature.
This causes issues such as the one reported by #3594 where
`EXECUTE_BATCH_SIZE` env var is not used.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Include p2p's `sync-test` feature in crate's `sync-test` feature
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3594