feat(l1): add specific error message when import is used with the wrong network (#3893)
**Motivation**
When using `import` subcommand with a chain.rlp file that doesn't match
the network, we issue the regular `Parent Block Not Found` error. We are
able to distinguish this specific case as it will mean that the parent
block of block number 1 (aka the genesis block) is not found. This PR
takes advantage of this to issue a specific, more informative warning
message for this specific case
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Add a separate warning message when the imported chain file is not
compatible with the genesis
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3813