chore(l2): clean makefile (#3870)
**Motivation**

Improve our Makefile is a tech debt we are already suffering from. It is
outdated, with unused or broken targets, and it's disorganized, making
it hard to find useful ones.
While reviewing it, I found it easier to make the changes myself rather
than creating issues.

**Description**

- Organizes the file into categories (l1,l2,prover,etc). Inspired in
#3373.
- Removes unused targets.
- Since #3604, we only need `clean-contract-deps`, as our tests still
clone and compile contracts. This PR removes the target and adds a step
at the end of the integration tests to clean those directories.
- Adds `deploy-l1-sp1` target to deploy the `sp1-verifier` contract,
useful for running tests on servers with a GPU.

Closes None

---------

Co-authored-by: Tomás Paradelo <112426153+tomip01@users.noreply.github.com>