feat(l2): ethrex replay add metrics & logs (#3785)
**Motivation**

Ethrex replay is not showing enough information which might lead users
to think the program is stuck when it's actually doing things. Also we
want to have a way to easily create samply profiles for sp1

**Description**

- Add `profiling` feature to ethrex-replay and ethrex-prover to enable
the profiling feature in sp1
- Add a makefile command to easily generate a samply for an l2 batch or
an l1 block
- Set tracing level to INFO for ethrex_replay, disable some sp1 logs
that were too spammy, add info! logs to ethrex replay
- Logs where the data for the prover is coming from (local cache or RPC
server)
   - Logs the amount of time an it took for the execution program to run
   - Logs the amount of gas that was proven/executed
- Fixes an issue where getting data for a batch from a validium mode L2
failed because when in validium mode the blobs field is empty

---------

Co-authored-by: Tomás Paradelo <112426153+tomip01@users.noreply.github.com>