fix(l1): saner behavior on low nonce error
During block building we check transaction validity inside the execution
step, only to go all-or-nothing with the result.
Because of the extreme position we take, we conservatively only skip the
tx for this block, trying again the next one in case the error was
recoverable.
When the error is caused by a transaction nonce being lower than
expected, this ensures it never resolves. A more reasonable approach is
catching that particular error, removing the transaction from the
mempool, and then skipping only that transaction for this block, rather
than all transactions from that sender.