ci(l1): fix Docker Login CI. (#3935)
**Motivation**
External contributors who forked the repo were experiencing CI failures
during Docker Hub login because they didn’t have access to our Docker
credentials (stored as repository variables and secrets)

This PR fixes the issue by adding a conditional step to the GitHub
Action workflow. It checks if Docker credentials are available, and if
they're not set (as in forks), the login step is skipped.