fix(l1): fix overlapping chunk ids (#3985)
**Motivation**

As part of snap sync, we download accounts which generates several
batches.

For each of these batches, storage download requests are made. Those are
numbered with a chunk_id.

For the last chunk of each storage download, the chunk_id isn't
incremented This causes the first chunk of the next batch to overwrite
the last chunk of the previous one.

This causes some storage ranges to be lost when more than one chunk is
generated.

**Description**

Fixes the off-by-one error.