fix: snap sync storage download metrics (#4026)
**Motivation**
The metrics for the storage download were confusing since the progress
shown was for each chunk of accounts, so the percentage would go from 0%
to 100% for each chunk, making it difficult to track the progress.

**Description**
To fix this, we calculate the number of storage tries that we need to
download, and we maintain a counter for the downloaded storages tries.
In each iteration, the `request_storage_ranges` updates the counter.

Then, we use the counter along with the total storages to calculate the
progress.

Closes #issue_number