fix(l1): fix off-by-one in update_pivot (#4030)
**Motivation**

When snap syncing, we may update the pivot. This function fetches the
extra block headers between the old and new pivot position.

However, it fetches an extra block (the old pivot, that we already
have). This causes the list of block hashes to have a duplicate item,
which causes errors when calculating the block numbers
(numbers_and_hashes).

**Description**

Fixes the off-by-one, uncomments the update_pivot call and ensures the
updated block hash list is passed to into_fullsync.

---------

Co-authored-by: Gianbelinche <39842759+gianbelinche@users.noreply.github.com>