chore(l2): avoid CI_ETHREX_WORKDIR docker warnings (#4064)
**Motivation**

Running `make down` or `make init` prints the following warnings:

```
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
WARN[0000] The "CI_ETHREX_WORKDIR" variable is not set. Defaulting to a blank string.
```

The same was happening for the `L1_RPC_URL` variable.

**Description**

- Sets the variables when running docker compose targets.
- Renames `CI_ETHREX_WORKDIR` to `DOCKER_ETHREX_WORKDIR`.

Closes #3810