log `InconsistentTree` as debug while syncing (#4084)
**Motivation**
While the node is syncing, we occasionally see warnings with the
`InconsistentTree` error.
This occurs because requests may ask for data that has not yet been
downloaded during sync.

Since this situation is expected and not truly an error, the log level
should be adjusted. However, if the same issue happens once the node is
fully synced, it indicates a real problem and should still be logged as
an error.

**Description**
This change updates the logging logic:
* When the node is fully synced: log `InconsistentTree` as `error`.
* When the node is syncing: log `InconsistentTree` as `debug`.