chore(l1): use `local` tag for ethrex images for local hive and kurtosis (#4132)
**Motivation**
Make it explicit that we're using a locally build version of ethrex.
Decouple CI from local use of docker images

**Description**
Since `latest` is the default tag if none is specified, we had bugs in
the past where we were accidentally using an old version of ethrex
instead of the locally build one. Lets make it explicit that we're using
a local version by using the `local` tag instead of `latest`.