refactor(l1,l2): rename genserver errors (#3825)
**Motivation**

End user might (and shouldn't) be aware of what a `GenServer`
represents, so seing an error of such type would be confusing.

**Description**

This PR rewords the _GenServer_ errors to _Internal_ such that they are
still as detailed as before, but without the technical wording. There
were some parts of the where `InternalError` already existed, in such
cases this types of errors have been expanded to different, new, types.

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #3376

---------

Co-authored-by: Tomás Paradelo <112426153+tomip01@users.noreply.github.com>