fix(l2): properly restore context in payload_builder (#4242)
**Motivation**

We have a bug in our `payload_builder` when filling transactions into a
block, which appears later during the block validation process of the
`prover`:


```
2025-09-01T18:27:49.421092Z ERROR ethrex_prover_lib::prover: Gas validation error: Invalid Block: Gas used doesn't match value in header. Used: 46011000, Expected: 50064000 endpoint=tcp://127.0.0.1:3900
```

#### What is it about?

After executing a transaction, if the block’s `stateDiff` size exceeds
the maximum allowed, we restore the `vm.context`, reverting the account
updates generated by the transaction.
The problem is that we should also update `context.remaining_gas` and
`context.block_value` as well.

#### Why haven’t we seen the error before?

There is an `if` check that prevents the error from appearing in our
daily usage:


```Rust
// Check if we have enough space for the StateDiff to run more transactions
if acc_size_without_accounts + size_accounts_diffs + SIMPLE_TX_STATE_DIFF_SIZE
> safe_bytes_per_blob {
	warn!("No more StateDiff space to run transactions");
	break;
};
```

This is an early return that checks if we have enough space for a simple
transfer, and stops adding transactions to the block if we don’t.
However, it can also happen that we have enough space for a simple
transfer, but the transaction we are trying to add occupies more space
than that, making us restore the previous state.

#### How to trigger the error

The easiest way to trigger the error is to comment out the early return,
allowing the `payload_builder` to attempt restoring the previous state.
Then start the `L2` and `prover` with:

```
make init
make init-prover
```

and run a load test with:
```
cargo run --release --manifest-path ./tooling/load_test/Cargo.toml -- -k ./fixtures/keys/private_keys_l1.txt -t eth-transfers -n "http://localhost:1729"
```


**Description**

This PR updates the `paylaod_builder` to restore the `remaining_gas` and
the `block_value` If the `stateDiff` size exceeds the maximum allowed
after executing a transaction.

Issue #4243 was created to add an integration test covering this case.

Closes None