feat(l1): improve p2p peer discovery and rework snap sync (#4213)
**Motivation**

This PR does two main things:

- Improves the process of finding peers by being way more aggressive in
our lookup. The previous version was too passive and as a result did not
find many peers in a reasonable amount of time.
- Reworks and fixes a few things in snap sync to make it work on the
main networks: this was tested and ran succesfully on hoodi, sepolia and
mainnet. It currently syncs hoodi in around 20 minutes in a good
scenario, sepolia in around 3 and a half hours, and mainnet in ~24
hours. These times are very far from ideal and there is a ton of room
for improvement here. This is a first iteration that will be improved in
subsequent PRs, along with improvements to our p2p code in general (lots
of bugs and performance issues were found once we started succesfully
syncing big testnets). The code is also far from ideal from a
readability standpoint, something we will also tackle on a future PR by
rewriting this code using `spawned`.

Some of the things left to tackle on subsequent PRs are:

- Fix memory usage, currently the snap sync code does not flush data to
disk frequently enough and as a result uses massive amounts of memory on
big networks when syncing.
- Use a global peer scoring mechanism instead of tracking scores
separately on each snap sync stage.
- Improve download task distribution to peers; the algorithm is
currently too simple and becomes very slow if run on a node with high
latency to most of its peers.
- Various p2p fixes: improve tx broadcast logic to decrease cpu usage
when we have lots of peers, limit mempool size and other changes needed
to make the node more stable.
- Optionally download blocks and receipts as well once the sync process
is finished. These are not necessary for a node to be synced but they
are common data that most clients usually serve by default.
- Various performance improvements to reduce time spent writing to the
database, make db writes concurrent to downloads, and more.

---------

Co-authored-by: Tomás Grüner <47506558+MegaRedHand@users.noreply.github.com>
Co-authored-by: ilitteri <ilitteri@fi.uba.ar>
Co-authored-by: Pablo Deymonnaz <deymonnaz@gmail.com>
Co-authored-by: Lucas Fiegl <iovoid@users.noreply.github.com>
Co-authored-by: ricomateo <mrico@fi.uba.ar>
Co-authored-by: Gianbelinche <39842759+gianbelinche@users.noreply.github.com>
Co-authored-by: Francisco Xavier Gauna <francisco.gauna@lambdaclass.com>
Co-authored-by: Mateo Rico <89949621+ricomateo@users.noreply.github.com>
Co-authored-by: Juan Munoz <juanmunoz890@gmail.com>
Co-authored-by: Esteban Dimitroff Hodi <esteban.dimitroff@lambdaclass.com>
Co-authored-by: Pablo Deymonnaz <pdeymon@fi.uba.ar>
Co-authored-by: juan518munoz <62400508+juan518munoz@users.noreply.github.com>