fix(levm): check coordinates are within field bounds in ecpairing precompile (#4082)
**Motivation**
A gas mismatch was found on a [sepolia
transaction](https://sepolia.etherscan.io/tx/0x4355d49be46e61a53c71f45a128ebefb52cb38df08ed55833c2c162d26396819).
This was due to the transaction failing at an ecpairing precompile on
the network but not on levm. The difference was that geth checks that
the coordinates received are within the bounds of the field, while we
perform modulus on the input coordinates (see
`BN254FieldElement::from_bytes_be`). This PR fixes this by checking that
the input coordinates are within the field bounds before constructing
the FieldElements.
It also adds some tests taken from geth to ensure that the
implementation remains correct + a test for the out of bounds case
<!-- Why does this pull request exist? What are its goals? -->

**Description**
* Check that input coordinates are within field bounds in `ecpairing`
precompile
* Add tests for `ecpairing` precompile
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes #issue_number