chore(l1): adjust bytecode chunk size (#4311)
**Motivation**

Bytecode download is currently slow, in part due to the small chunk
size.

**Description**

Considering that the maximum size of a contract is 0x6000, the new limit
would mean a max usage of ~1.2GB. In practice it's even lower because
not all contracts are that big, and many accounts point to the same
contracts.