feat(l1): improve transaction broadcasting logic (#4259)
**Motivation**

Current version just broadcast every transaction we receive to every
active peer we have. This presents some issues on performance, memory
and even in code responsibilities.
<!-- Why does this pull request exist? What are its goals? -->

**Description**
Reorganizes the transaction broadcast logic
When a new peer connection is initialized, we send the hashes of all the
txs in the mempool.
Then every time we receive new txs, we add them to the mempool and send
them to our peers, we send the txs to a sqrt of our peers, while sending
only the hashes to the rest.
The mempool itself is the one responsible for knowing which txs need to
be broadcasted. Every time we add a new tx to the mempool we add it to
the list of pending broadcast, then once we broadcast them, we erase
them. Also if a tx is removed from the mempool before being broadcasted
it is also deleted from the broadcast pool.

There is one possible scenario where we could send a tx twice to the
same peer, that is if after the tx left the mempool, we receive it again
and add it back, we would send it again to the peers.
<!-- A clear and concise general description of the changes this PR
introduces -->

<!-- Link to issues: Resolves #111, Resolves #222 -->

Closes https://github.com/lambdaclass/ethrex/issues/4241