fix(l1): replay witness get code missing (#4280)
**Motivation**

<!-- Why does this pull request exist? What are its goals? -->

**Description**

<!-- A clear and concise general description of the changes this PR
introduces -->
- Gets the witness with the proper block number instead of using Latest
(when it wasn't specified) to avoid getting a witness that differs with
the requested block number that we say we are looking for.
- Different execution clients access state at different times, one can
access code of an account before a particular validation and the other
one can access it later. In the case in which the validation failed, the
latter client will generate a witness without the code whereas the
former will actually need that code.
- A possible fix would be to access data ONLY when we are 100% sure that
we are going to use it. But this would mean making the code uglier in
order to fix this, which is probably not worth it.
- The fix that I propose here is to assume that the witness is correct
and return a default value when it wasn't found in it, because it would
mean that it isn't relevant for execution. We also print a warning
alongside this for debugging purposes if there's a state root mismatch.