fix(l1): prevent race conditions on getReceipts calls (#4346)
**Motivation**

We are encountering the following error in our L2 integration tests:
```
2025-08-29T19:17:04.573865Z  WARN ethrex_l2::sequencer::l1_watcher: Error when getting logs from L1: eth_getLogs request error: Internal Error: Could not get receipt
```

After debugging, it turned out to be a race condition that can occur if
a new block has just been marked as canonical in the store cache, but
has not yet been updated in the db. If a `getReceipt` call arrives in
this window, the node won't be able to find the receipts by block
number.

**Description**

- Updates the engine API to identify blocks by hash instead of block
number.
- Resolves the block hash by calling `get_canonical_block_hash` in the
store.


Closes None