fix(l2): update privileged_range once all checks have passed (#4224)
**Motivation**

Noticed a weird behavior while testing #4211. Once the configured
`max_gas_limit` was reached, no further deposits were included in
subsequent blocks until a new batch is started. This happens because we
first update the `privileged_range`, and if the transaction is later
rejected (for example, due to the gas limit), we don’t remove the nonce
from the range. As a result, when the next block starts, we try to add
the tx again, but since the nonce was already included to the range, the
transaction is marked as out-of-order.

#### Why haven’t we seen the error before?

There is an `if` check that prevents the error from appearing in our
daily usage:

```Rust
// Check if we have enough space for the StateDiff to run more transactions
if acc_size_without_accounts + size_accounts_diffs + SIMPLE_TX_STATE_DIFF_SIZE
> safe_bytes_per_blob {
	warn!("No more StateDiff space to run transactions");
	break;
};
```

This is an early return that checks if we have enough space for a simple
transfer in our `stateDiff`, and stops adding transactions to the block
if we don’t. But, it can happen that we have enough space for a simple
transfer, but the transaction we are trying to add occupies more space
than that, making us restore the previous state in another check a[ few
lines
later](https://github.com/lambdaclass/ethrex/blob/653a430e8a15d624a36c8ecef1c0fc156af0dae0/crates/l2/sequencer/block_producer/payload_builder.rs#L215).

<!-- 
#### How to trigger the error

The easiest way to trigger the error is to comment out the early return,
allowing the `payload_builder` to attempt restoring the previous state.
Then start the `L2` and `prover` with:

```
make init
make init-prover
```

and run a load test with:
```
cargo run --release --manifest-path ./tooling/load_test/Cargo.toml -- -k ./fixtures/keys/private_keys_l1.txt -t eth-transfers -n "http://localhost:1729"
```

Observe the following behavior in the monitor:

After a block reaches its max `stateDiff` size allowed, the subsequent
blocks contain 0 transactions until a new batch is started.
-->

**Description**

Moves the update of `privileged_range` after all other checks have
passed.


Issues #4248 and #4243 were created.
Closes None